chore(pgwire): remove legacy PGWire server code
