+++ core/src/main/java/io/questdb/DefaultFactoryProvider.java
@@ -30,9 +30,14 @@ import io.questdb.cairo.security.AllowAllSecurityContextFactory;
 import io.questdb.cairo.security.SecurityContextFactory;
 import io.questdb.cutlass.auth.DefaultLineAuthenticatorFactory;
 import io.questdb.cutlass.auth.LineAuthenticatorFactory;
-import io.questdb.cutlass.http.*;
+import io.questdb.cutlass.http.DefaultHttpAuthenticatorFactory;
+import io.questdb.cutlass.http.DefaultHttpCookieHandler;
+import io.questdb.cutlass.http.DefaultHttpHeaderParserFactory;
+import io.questdb.cutlass.http.HttpAuthenticatorFactory;
+import io.questdb.cutlass.http.HttpCookieHandler;
+import io.questdb.cutlass.http.HttpHeaderParserFactory;
+import io.questdb.cutlass.pgwire.DefaultPgWireAuthenticatorFactoryModern;
 import io.questdb.cutlass.pgwire.PgWireAuthenticatorFactory;
-import io.questdb.cutlass.pgwire.modern.DefaultPgWireAuthenticatorFactoryModern;
 import io.questdb.network.PlainSocketFactory;
 import io.questdb.network.SocketFactory;
 import org.jetbrains.annotations.NotNull;
