+++ core/src/main/java/io/questdb/FactoryProviderImpl.java
@@ -32,8 +32,8 @@ import io.questdb.cutlass.http.DefaultHttpCookieHandler;
 import io.questdb.cutlass.http.DefaultHttpHeaderParserFactory;
 import io.questdb.cutlass.http.HttpAuthenticatorFactory;
 import io.questdb.cutlass.http.HttpCookieHandler;
+import io.questdb.cutlass.pgwire.DefaultPgWireAuthenticatorFactoryModern;
 import io.questdb.cutlass.pgwire.PgWireAuthenticatorFactory;
-import io.questdb.cutlass.pgwire.modern.DefaultPgWireAuthenticatorFactoryModern;
 import io.questdb.network.PlainSocketFactory;
 import io.questdb.network.SocketFactory;
 import org.jetbrains.annotations.NotNull;
