+++ core/src/main/java/io/questdb/PropertyKey.java
@@ -412,10 +412,8 @@ public enum PropertyKey implements ConfigPropertyKey {
     PG_INSERT_CACHE_BLOCK_COUNT("pg.insert.cache.block.count"),
     PG_INSERT_CACHE_ROW_COUNT("pg.insert.cache.row.count"),
     PG_INSERT_POOL_CAPACITY("pg.insert.pool.capacity"),
-    PG_LEGACY_MODE_ENABLED("pg.legacy.mode.enabled"),
     PG_NAMED_STATEMENT_LIMIT("pg.named.statement.limit"),
     PG_NAMED_STATEMENT_CACHE_CAPACITY("pg.named.statement.cache.capacity"),
-    PG_NAMED_STATEMENT_POOL_CAPACITY("pg.named.statement.pool.capacity"),
     PG_PENDING_WRITERS_CACHE_CAPACITY("pg.pending.writers.cache.capacity"),
     PG_NET_CONNECTION_SNDBUF("pg.net.connection.sndbuf"),
     DEBUG_FORCE_SEND_FRAGMENTATION_CHUNK_SIZE("debug.force.send.fragmentation.chunk.size", false, true),
