+++ core/src/main/java/io/questdb/cutlass/pgwire/PGResumeCallback.java
@@ -22,10 +22,9 @@
  *
  ******************************************************************************/
 
-package io.questdb.cutlass.pgwire.modern;
+package io.questdb.cutlass.pgwire;
 
-import io.questdb.std.IntList;
-
-public interface TypeContainer {
-    IntList getPgInParameterTypeOIDs();
+@FunctionalInterface
+interface PGResumeCallback {
+    void resume() throws Exception;
 }
