+++ core/src/main/java/io/questdb/cutlass/pgwire/TypesAndSelectModern.java
@@ -22,7 +22,7 @@
  *
  ******************************************************************************/
 
-package io.questdb.cutlass.pgwire.modern;
+package io.questdb.cutlass.pgwire;
 
 import io.questdb.cairo.sql.RecordCursorFactory;
 import io.questdb.std.IntList;
