+++ core/src/test/java/io/questdb/test/cutlass/pgwire/BasePGTest.java
@@ -54,10 +54,8 @@ import io.questdb.std.str.Utf16Sink;
 import io.questdb.test.AbstractCairoTest;
 import io.questdb.test.mp.TestWorkerPool;
 import io.questdb.test.tools.TestUtils;
-import org.checkerframework.checker.nullness.qual.NonNull;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
-import org.junit.Assume;
 import org.postgresql.util.PSQLException;
 
 import java.io.IOException;
@@ -71,8 +69,6 @@ import java.sql.ResultSet;
 import java.sql.ResultSetMetaData;
 import java.sql.SQLException;
 import java.sql.Timestamp;
-import java.util.Arrays;
-import java.util.Collection;
 import java.util.Properties;
 import java.util.TimeZone;
 
@@ -89,7 +85,6 @@ public abstract class BasePGTest extends AbstractCairoTest {
     public static final long CONN_AWARE_EXTENDED = CONN_AWARE_EXTENDED_LIMITED | CONN_AWARE_QUIRKS;
     public static final long CONN_AWARE_SIMPLE = 2;
     public static final long CONN_AWARE_ALL = CONN_AWARE_SIMPLE | CONN_AWARE_EXTENDED;
-    protected final boolean legacyMode;
     protected CopyRequestJob copyRequestJob = null;
     protected int forceRecvFragmentationChunkSize = 1024 * 1024;
     protected int forceSendFragmentationChunkSize = 1024 * 1024;
@@ -97,10 +92,6 @@ public abstract class BasePGTest extends AbstractCairoTest {
     protected int recvBufferSize = 1024 * 1024;
     protected int selectCacheBlockCount = -1;
     protected int sendBufferSize = 1024 * 1024;
-
-    protected BasePGTest(@NonNull LegacyMode legacyMode) {
-        this.legacyMode = legacyMode == LegacyMode.LEGACY;
-    }
 
     public static void assertResultSet(CharSequence expected, StringSink sink, ResultSet rs) throws SQLException, IOException {
         assertResultSet(null, expected, sink, rs);
@@ -281,14 +272,6 @@ public abstract class BasePGTest extends AbstractCairoTest {
         return rows;
     }
 
-    public void skipInLegacyMode() {
-        Assume.assumeFalse("Test does not support legacy mode", legacyMode);
-    }
-
-    public void skipInModernMode() {
-        Assume.assumeTrue("Test does not support modern mode", legacyMode);
-    }
-
     private static void toSink(InputStream is, Utf16Sink sink) throws IOException {
         // limit what we print
         byte[] bb = new byte[1];
@@ -387,12 +370,6 @@ public abstract class BasePGTest extends AbstractCairoTest {
         }
     }
 
-    static Collection<Object[]> legacyModeParams() {
-        return Arrays.asList(new Object[][]{
-                {LegacyMode.MODERN}, {LegacyMode.LEGACY}
-        });
-    }
-
     protected void assertWithPgServer(
             Mode mode,
             boolean binary,
@@ -455,9 +432,6 @@ public abstract class BasePGTest extends AbstractCairoTest {
             PGWireConfiguration configuration,
             boolean fixedClientIdAndSecret
     ) throws SqlException {
-        if (configuration.isLegacyModeEnabled() != legacyMode) {
-            ((Port0PGWireConfiguration) configuration).isLegacyMode = legacyMode;
-        }
         TestWorkerPool workerPool = new TestWorkerPool(configuration);
         copyRequestJob = new CopyRequestJob(engine, configuration.getWorkerCount());
 
@@ -502,7 +476,7 @@ public abstract class BasePGTest extends AbstractCairoTest {
             }
         };
 
-        final PGWireConfiguration conf = new Port0PGWireConfiguration(-1, legacyMode) {
+        final PGWireConfiguration conf = new Port0PGWireConfiguration(-1) {
 
             @Override
             public SqlExecutionCircuitBreakerConfiguration getCircuitBreakerConfiguration() {
@@ -623,11 +597,6 @@ public abstract class BasePGTest extends AbstractCairoTest {
             public Rnd getRandom() {
                 return new Rnd();
             }
-
-            @Override
-            public boolean isLegacyModeEnabled() {
-                return legacyMode;
-            }
         };
     }
 
@@ -653,16 +622,7 @@ public abstract class BasePGTest extends AbstractCairoTest {
             public Rnd getRandom() {
                 return new Rnd();
             }
-
-            @Override
-            public boolean isLegacyModeEnabled() {
-                return legacyMode;
-            }
         };
-    }
-
-    public enum LegacyMode {
-        MODERN, LEGACY
     }
 
     public enum Mode {
