+++ core/src/test/java/io/questdb/test/cutlass/pgwire/PGDeclareSyntaxTest.java
@@ -24,30 +24,12 @@
 
 package io.questdb.test.cutlass.pgwire;
 
-import org.checkerframework.checker.nullness.qual.NonNull;
 import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.junit.runners.Parameterized;
 
 import java.sql.PreparedStatement;
 import java.sql.ResultSet;
-import java.util.Arrays;
-import java.util.Collection;
 
-@RunWith(Parameterized.class)
 public class PGDeclareSyntaxTest extends BasePGTest {
-
-    public PGDeclareSyntaxTest(@NonNull LegacyMode legacyMode) {
-        super(legacyMode);
-    }
-
-    @Parameterized.Parameters(name = "{0}, {1}")
-    public static Collection<Object[]> testParams() {
-        return Arrays.asList(new Object[][]{
-                {LegacyMode.MODERN},
-                {LegacyMode.LEGACY},
-        });
-    }
 
     @Test
     public void testDeclareSyntaxWorksWithPositionalBindVariables() throws Exception {
@@ -69,7 +51,6 @@ public class PGDeclareSyntaxTest extends BasePGTest {
 
     @Test
     public void testDeclareSyntaxWorksWithPositionalBindVariables2() throws Exception {
-        skipInLegacyMode();
         assertWithPgServer(CONN_AWARE_ALL, (connection, binary, mode, port) -> {
             try (PreparedStatement ps = connection.prepareStatement("DECLARE @x := ? SELECT x as x2, (@x + 5) as x1, @x as x3 FROM (SELECT * FROM long_sequence(10) WHERE x = @x)")) {
                 ps.setInt(1, 5);
