+++ core/src/test/java/io/questdb/test/cutlass/pgwire/PGJobContextTest.java
@@ -189,8 +189,7 @@ public class PGJobContextTest extends BasePGTest {
     private final Rnd bufferSizeRnd = TestUtils.generateRandom(LOG);
     private final boolean walEnabled;
 
-    public PGJobContextTest(LegacyMode legacyMode, WalMode walMode) {
-        super(legacyMode);
+    public PGJobContextTest(WalMode walMode) {
         this.walEnabled = (walMode == WalMode.WITH_WAL);
     }
 
@@ -212,14 +211,9 @@ public class PGJobContextTest extends BasePGTest {
         stringTypeName = ColumnType.nameOf(ColumnType.STRING);
     }
 
-    @Parameters(name = "{0}, {1}")
-    public static Collection<Object[]> testParams() {
-        return Arrays.asList(new Object[][]{
-                {LegacyMode.MODERN, WalMode.WITH_WAL},
-                {LegacyMode.MODERN, WalMode.NO_WAL},
-                {LegacyMode.LEGACY, WalMode.WITH_WAL},
-                {LegacyMode.LEGACY, WalMode.NO_WAL},
-        });
+    @Parameters(name = "{0}")
+    public static Collection<Object> testParams() {
+        return List.of(new Object[]{WalMode.WITH_WAL, WalMode.NO_WAL});
     }
 
     @Before
@@ -2242,21 +2236,11 @@ if __name__ == "__main__":
             try (PreparedStatement ps = connection.prepareStatement("tab1 where ? is not null")) {
                 ps.setInt(1, Numbers.INT_NULL);
                 try (ResultSet rs = ps.executeQuery()) {
-                    if (binary || !legacyMode) {
-                        assertResultSet(
-                                "value[INTEGER],ts[TIMESTAMP]\n",
-                                sink,
-                                rs
-                        );
-                    } else {
-                        assertResultSet(
-                                "value[INTEGER],ts[TIMESTAMP]\n" +
-                                        "100,1970-01-01 00:00:00.0\n" +
-                                        "null,1970-01-01 00:00:00.000001\n",
-                                sink,
-                                rs
-                        );
-                    }
+                    assertResultSet(
+                            "value[INTEGER],ts[TIMESTAMP]\n",
+                            sink,
+                            rs
+                    );
                 }
             }
 
@@ -2274,49 +2258,27 @@ if __name__ == "__main__":
                 }
             }
 
-            if (legacyMode) {
-                // this would be an error on old PGWire
-                try (PreparedStatement ps = connection.prepareStatement("tab1 where ? is not null")) {
-                    ps.setString(1, "");
-                    try (ResultSet ignore1 = ps.executeQuery()) {
-                        Assert.fail();
-                    } catch (PSQLException e) {
-                        TestUtils.assertContains(e.getMessage(), "inconvertible value: `` [" + stringTypeName + " -> DOUBLE]");
-                    }
+            // not an error in PGWire 2.0
+            try (PreparedStatement ps = connection.prepareStatement("tab1 where ? is not null")) {
+                ps.setString(1, "");
+                try (ResultSet rs = ps.executeQuery()) {
+                    assertResultSet(
+                            "value[INTEGER],ts[TIMESTAMP]\n" +
+                                    "100,1970-01-01 00:00:00.0\n" +
+                                    "null,1970-01-01 00:00:00.000001\n",
+                            sink,
+                            rs
+                    );
                 }
-
-                try (PreparedStatement ps = connection.prepareStatement("tab1 where ? is not null")) {
-                    ps.setString(1, "cah-cha-cha");
-                    try (ResultSet ignore1 = ps.executeQuery()) {
-                        Assert.fail();
-                    } catch (PSQLException e) {
-                        TestUtils.assertContains(e.getMessage(), "inconvertible value: `cah-cha-cha` [" + stringTypeName + " -> DOUBLE]");
-                    }
-                }
-
-            } else {
-                // not an error in PGWire 2.0
-                try (PreparedStatement ps = connection.prepareStatement("tab1 where ? is not null")) {
-                    ps.setString(1, "");
-                    try (ResultSet rs = ps.executeQuery()) {
-                        assertResultSet(
-                                "value[INTEGER],ts[TIMESTAMP]\n" +
-                                        "100,1970-01-01 00:00:00.0\n" +
-                                        "null,1970-01-01 00:00:00.000001\n",
-                                sink,
-                                rs
-                        );
-                    }
-                    ps.setString(1, "cah-cha-cha");
-                    try (ResultSet rs = ps.executeQuery()) {
-                        assertResultSet(
-                                "value[INTEGER],ts[TIMESTAMP]\n" +
-                                        "100,1970-01-01 00:00:00.0\n" +
-                                        "null,1970-01-01 00:00:00.000001\n",
-                                sink,
-                                rs
-                        );
-                    }
+                ps.setString(1, "cah-cha-cha");
+                try (ResultSet rs = ps.executeQuery()) {
+                    assertResultSet(
+                            "value[INTEGER],ts[TIMESTAMP]\n" +
+                                    "100,1970-01-01 00:00:00.0\n" +
+                                    "null,1970-01-01 00:00:00.000001\n",
+                            sink,
+                            rs
+                    );
                 }
             }
 
@@ -2420,23 +2382,14 @@ if __name__ == "__main__":
                 // server cannot assume that the client is sending null
                 ps.setInt(1, Numbers.INT_NULL);
                 try (ResultSet rs = ps.executeQuery()) {
-                    if (binary || !legacyMode) {
-                        // in binary protocol DOUBLE.null == INT.null
-                        assertResultSet(
-                                "value[INTEGER],ts[TIMESTAMP]\n" +
-                                        "100,1970-01-01 00:00:00.0\n" +
-                                        "null,1970-01-01 00:00:00.000001\n",
-                                sink,
-                                rs
-                        );
-                    } else {
-                        // in string protocol DOUBLE.null != INT.null
-                        assertResultSet(
-                                "value[INTEGER],ts[TIMESTAMP]\n",
-                                sink,
-                                rs
-                        );
-                    }
+                    // in binary protocol DOUBLE.null == INT.null
+                    assertResultSet(
+                            "value[INTEGER],ts[TIMESTAMP]\n" +
+                                    "100,1970-01-01 00:00:00.0\n" +
+                                    "null,1970-01-01 00:00:00.000001\n",
+                            sink,
+                            rs
+                    );
                 }
             }
 
@@ -2455,42 +2408,22 @@ if __name__ == "__main__":
                 }
             }
 
-            if (legacyMode) {
-                try (PreparedStatement ps = connection.prepareStatement("tab1 where ? is null")) {
-                    ps.setString(1, "");
-                    try (ResultSet ignore1 = ps.executeQuery()) {
-                        Assert.fail();
-                    } catch (PSQLException e) {
-                        TestUtils.assertContains(e.getMessage(), "inconvertible value: `` [" + stringTypeName + " -> DOUBLE]");
-                    }
+            try (PreparedStatement ps = connection.prepareStatement("tab1 where ? is null")) {
+                ps.setString(1, "");
+                try (ResultSet rs = ps.executeQuery()) {
+                    assertResultSet(
+                            "value[INTEGER],ts[TIMESTAMP]\n",
+                            sink,
+                            rs
+                    );
                 }
-
-                try (PreparedStatement ps = connection.prepareStatement("tab1 where ? is null")) {
-                    ps.setString(1, "cha-cha-cha");
-                    try (ResultSet ignore1 = ps.executeQuery()) {
-                        Assert.fail();
-                    } catch (PSQLException e) {
-                        TestUtils.assertContains(e.getMessage(), "inconvertible value: `cha-cha-cha` [" + stringTypeName + " -> DOUBLE]");
-                    }
-                }
-            } else {
-                try (PreparedStatement ps = connection.prepareStatement("tab1 where ? is null")) {
-                    ps.setString(1, "");
-                    try (ResultSet rs = ps.executeQuery()) {
-                        assertResultSet(
-                                "value[INTEGER],ts[TIMESTAMP]\n",
-                                sink,
-                                rs
-                        );
-                    }
-                    ps.setString(1, "cha-cha-cha");
-                    try (ResultSet rs = ps.executeQuery()) {
-                        assertResultSet(
-                                "value[INTEGER],ts[TIMESTAMP]\n",
-                                sink,
-                                rs
-                        );
-                    }
+                ps.setString(1, "cha-cha-cha");
+                try (ResultSet rs = ps.executeQuery()) {
+                    assertResultSet(
+                            "value[INTEGER],ts[TIMESTAMP]\n",
+                            sink,
+                            rs
+                    );
                 }
             }
 
@@ -2540,8 +2473,6 @@ if __name__ == "__main__":
 
     @Test
     public void testBindVariablesWithNonIndexedSymbolInFilter() throws Exception {
-        // @Ignore("ERROR: bind variable at 0 is defined as unknown and cannot accept STRING")
-        Assume.assumeFalse(legacyMode);
         testBindVariablesWithIndexedSymbolInFilter(false);
     }
 
@@ -2595,35 +2526,21 @@ if __name__ == "__main__":
     public void testBrokenUtf8QueryInParseMessage() throws Exception {
         skipOnWalRun(); // non-partitioned
         final String script;
-        if (legacyMode) {
-            // legacy mode will disconnect if UTF8 string is bad
-            script = ">0000000804d2162f\n" +
-                    "<4e\n" +
-                    ">0000007500030000757365720061646d696e006461746162617365006e6162755f61707000636c69656e745f656e636f64696e67005554463800446174655374796c650049534f0054696d655a6f6e65004575726f70652f4c6f6e646f6e0065787472615f666c6f61745f64696769747300320000\n" +
-                    "<520000000800000003\n" +
-                    ">700000000a717565737400\n" +
-                    "<520000000800000000530000001154696d655a6f6e6500474d5400530000001d6170706c69636174696f6e5f6e616d6500517565737444420053000000187365727665725f76657273696f6e0031312e33005300000019696e74656765725f6461746574696d6573006f6e005300000019636c69656e745f656e636f64696e670055544638004b0000000c0000003fbb8b96505a0000000549\n" +
-                    ">50000000220053ac542065787472615f666c6f61745f646967697473203d2033000000420000000c0000000000000000450000000900000000015300000004\n" +
-                    "<!!";
-        } else {
-            // the modern server will return actionable error message to the client
-            script = ">0000000804d2162f\n" +
-                    "<4e\n" +
-                    ">0000007500030000757365720061646d696e006461746162617365006e6162755f61707000636c69656e745f656e636f64696e67005554463800446174655374796c650049534f0054696d655a6f6e65004575726f70652f4c6f6e646f6e0065787472615f666c6f61745f64696769747300320000\n" +
-                    "<520000000800000003\n" +
-                    ">700000000a717565737400\n" +
-                    "<520000000800000000530000001154696d655a6f6e6500474d5400530000001d6170706c69636174696f6e5f6e616d6500517565737444420053000000187365727665725f76657273696f6e0031312e33005300000019696e74656765725f6461746574696d6573006f6e005300000019636c69656e745f656e636f64696e670055544638004b0000000c0000003fbb8b96505a0000000549\n" +
-                    ">50000000220053ac542065787472615f666c6f61745f646967697473203d2033000000420000000c0000000000000000450000000900000000015300000004\n" +
-                    "<4500000039433030303030004d696e76616c6964205554463820627974657320696e20706172736520717565727900534552524f5200503100005a0000000549\n";
-        }
+        // the modern server will return actionable error message to the client
+        script = ">0000000804d2162f\n" +
+                "<4e\n" +
+                ">0000007500030000757365720061646d696e006461746162617365006e6162755f61707000636c69656e745f656e636f64696e67005554463800446174655374796c650049534f0054696d655a6f6e65004575726f70652f4c6f6e646f6e0065787472615f666c6f61745f64696769747300320000\n" +
+                "<520000000800000003\n" +
+                ">700000000a717565737400\n" +
+                "<520000000800000000530000001154696d655a6f6e6500474d5400530000001d6170706c69636174696f6e5f6e616d6500517565737444420053000000187365727665725f76657273696f6e0031312e33005300000019696e74656765725f6461746574696d6573006f6e005300000019636c69656e745f656e636f64696e670055544638004b0000000c0000003fbb8b96505a0000000549\n" +
+                ">50000000220053ac542065787472615f666c6f61745f646967697473203d2033000000420000000c0000000000000000450000000900000000015300000004\n" +
+                "<4500000039433030303030004d696e76616c6964205554463820627974657320696e20706172736520717565727900534552524f5200503100005a0000000549\n";
         assertHexScript(script);
     }
 
     @Test
     public void testByteBindingVariable() throws Exception {
         skipOnWalRun();
-        skipInLegacyMode();
-
         assertWithPgServer(CONN_AWARE_ALL, (connection, binary, mode, port) -> {
             final PreparedStatement statement = connection.prepareStatement("create table x (a byte)");
             statement.execute();
@@ -2772,8 +2689,6 @@ if __name__ == "__main__":
 
     @Test
     public void testCancelRunningQuery() throws Exception {
-        // legacy code is liable to "resume" cursor after "cursor.close()", which would lead to undefined behaviour
-        Assume.assumeFalse(legacyMode);
         String[] queries = {
                 "create table new_tab as (select count(*) from tab t1 join tab t2 on t1.x = t2.x where sleep(120000))",
                 "select count(*) from tab t1 join tab t2 on t1.x = t2.x where sleep(120000)",
@@ -2889,63 +2804,32 @@ if __name__ == "__main__":
     public void testCloseMessageForSelectWithParamsHex() throws Exception {
         skipOnWalRun();
         final String script;
-        if (legacyMode) {
-            // old behaviour:
-            //   - the client supplies parameter types, server ignores them and sends VARCHAR in the 'T' message
-            script = ">0000006e00030000757365720078797a0064617461626173650071646200636c69656e745f656e636f64696e67005554463800446174655374796c650049534f0054696d655a6f6e65004575726f70652f4c6f6e646f6e0065787472615f666c6f61745f64696769747300320000\n" +
-                    "<520000000800000003\n" +
-                    ">70000000076f6800\n" +
-                    "<520000000800000000530000001154696d655a6f6e6500474d5400530000001d6170706c69636174696f6e5f6e616d6500517565737444420053000000187365727665725f76657273696f6e0031312e33005300000019696e74656765725f6461746574696d6573006f6e005300000019636c69656e745f656e636f64696e670055544638004b0000000c0000003fbb8b96505a0000000549\n" +
-                    ">5000000022005345542065787472615f666c6f61745f646967697473203d2033000000420000000c0000000000000000450000000900000000015300000004\n" +
-                    "<310000000432000000044300000008534554005a0000000549\n" +
-                    ">500000003700534554206170706c69636174696f6e5f6e616d65203d2027506f737467726553514c204a4442432044726976657227000000420000000c0000000000000000450000000900000000015300000004\n" +
-                    "<310000000432000000044300000008534554005a0000000549\n" +
-                    ">500000003b0073656c65637420782c24312c24322c24332066726f6d206c6f6e675f73657175656e63652832290000030000001700000014000002bd420000002600000003000000000000000300000001340000000331323300000004352e3433000044000000065000450000000900000000005300000004\n" +
-                    "<31000000043200000004540000005900047800000000000001000000140008ffffffff000024310000000000000200000413ffffffffffff000024320000000000000300000413ffffffffffff000024330000000000000400000413ffffffffffff0000440000001f0004000000013100000001340000000331323300000004352e3433440000001f0004000000013200000001340000000331323300000004352e3433430000000d53454c4543542032005a0000000549\n" +
-                    ">500000003b0073656c65637420782c24312c24322c24332066726f6d206c6f6e675f73657175656e63652832290000030000001700000014000002bd420000002600000003000000000000000300000001340000000331323300000004352e3433000044000000065000450000000900000000005300000004\n" +
-                    "<31000000043200000004540000005900047800000000000001000000140008ffffffff000024310000000000000200000413ffffffffffff000024320000000000000300000413ffffffffffff000024330000000000000400000413ffffffffffff0000440000001f0004000000013100000001340000000331323300000004352e3433440000001f0004000000013200000001340000000331323300000004352e3433430000000d53454c4543542032005a0000000549\n" +
-                    ">500000003b0073656c65637420782c24312c24322c24332066726f6d206c6f6e675f73657175656e63652832290000030000001700000014000002bd420000002600000003000000000000000300000001340000000331323300000004352e3433000044000000065000450000000900000000005300000004\n" +
-                    "<31000000043200000004540000005900047800000000000001000000140008ffffffff000024310000000000000200000413ffffffffffff000024320000000000000300000413ffffffffffff000024330000000000000400000413ffffffffffff0000440000001f0004000000013100000001340000000331323300000004352e3433440000001f0004000000013200000001340000000331323300000004352e3433430000000d53454c4543542032005a0000000549\n" +
-                    ">500000003b0073656c65637420782c24312c24322c24332066726f6d206c6f6e675f73657175656e63652832290000030000001700000014000002bd420000002600000003000000000000000300000001340000000331323300000004352e3433000044000000065000450000000900000000005300000004\n" +
-                    "<31000000043200000004540000005900047800000000000001000000140008ffffffff000024310000000000000200000413ffffffffffff000024320000000000000300000413ffffffffffff000024330000000000000400000413ffffffffffff0000440000001f0004000000013100000001340000000331323300000004352e3433440000001f0004000000013200000001340000000331323300000004352e3433430000000d53454c4543542032005a0000000549\n" +
-                    ">500000003e535f310073656c65637420782c24312c24322c24332066726f6d206c6f6e675f73657175656e63652832290000030000001700000014000002bd420000002900535f31000003000000000000000300000001340000000331323300000004352e3433000044000000065000450000000900000000005300000004\n" +
-                    "<31000000043200000004540000005900047800000000000001000000140008ffffffff000024310000000000000200000413ffffffffffff000024320000000000000300000413ffffffffffff000024330000000000000400000413ffffffffffff0000440000001f0004000000013100000001340000000331323300000004352e3433440000001f0004000000013200000001340000000331323300000004352e3433430000000d53454c4543542032005a0000000549\n" +
-                    ">420000002900535f31000003000000000000000300000001340000000331323300000004352e34330000450000000900000000005300000004\n" +
-                    "<3200000004440000001f0004000000013100000001340000000331323300000004352e3433440000001f0004000000013200000001340000000331323300000004352e3433430000000d53454c4543542032005a0000000549\n" +
-                    ">430000000953535f31005300000004\n" +
-                    "<33000000045a0000000549\n" +
-                    ">5800000004\n";
-
-        } else {
-            // new bahviour:
-            //   - the server respects the parameter types sent by the client
-            script = ">0000006e00030000757365720078797a0064617461626173650071646200636c69656e745f656e636f64696e67005554463800446174655374796c650049534f0054696d655a6f6e65004575726f70652f4c6f6e646f6e0065787472615f666c6f61745f64696769747300320000\n" +
-                    "<520000000800000003\n" +
-                    ">70000000076f6800\n" +
-                    "<520000000800000000530000001154696d655a6f6e6500474d5400530000001d6170706c69636174696f6e5f6e616d6500517565737444420053000000187365727665725f76657273696f6e0031312e33005300000019696e74656765725f6461746574696d6573006f6e005300000019636c69656e745f656e636f64696e670055544638004b0000000c0000003fbb8b96505a0000000549\n" +
-                    ">5000000022005345542065787472615f666c6f61745f646967697473203d2033000000420000000c0000000000000000450000000900000000015300000004\n" +
-                    "<310000000432000000044300000008534554005a0000000549\n" +
-                    ">500000003700534554206170706c69636174696f6e5f6e616d65203d2027506f737467726553514c204a4442432044726976657227000000420000000c0000000000000000450000000900000000015300000004\n" +
-                    "<310000000432000000044300000008534554005a0000000549\n" +
-                    ">500000003b0073656c65637420782c24312c24322c24332066726f6d206c6f6e675f73657175656e63652832290000030000001700000014000002bd420000002600000003000000000000000300000001340000000331323300000004352e3433000044000000065000450000000900000000005300000004\n" +
-                    "<31000000043200000004540000005900047800000000000001000000140008ffffffff0000243100000000000002000000170004ffffffff0000243200000000000003000000140008ffffffff0000243300000000000004000002bd0008ffffffff0000440000001f0004000000013100000001340000000331323300000004352e3433440000001f0004000000013200000001340000000331323300000004352e3433430000000d53454c4543542032005a0000000549\n" +
-                    ">500000003b0073656c65637420782c24312c24322c24332066726f6d206c6f6e675f73657175656e63652832290000030000001700000014000002bd420000002600000003000000000000000300000001340000000331323300000004352e3433000044000000065000450000000900000000005300000004\n" +
-                    "<31000000043200000004540000005900047800000000000001000000140008ffffffff0000243100000000000002000000170004ffffffff0000243200000000000003000000140008ffffffff0000243300000000000004000002bd0008ffffffff0000440000001f0004000000013100000001340000000331323300000004352e3433440000001f0004000000013200000001340000000331323300000004352e3433430000000d53454c4543542032005a0000000549\n" +
-                    ">500000003b0073656c65637420782c24312c24322c24332066726f6d206c6f6e675f73657175656e63652832290000030000001700000014000002bd420000002600000003000000000000000300000001340000000331323300000004352e3433000044000000065000450000000900000000005300000004\n" +
-                    "<31000000043200000004540000005900047800000000000001000000140008ffffffff0000243100000000000002000000170004ffffffff0000243200000000000003000000140008ffffffff0000243300000000000004000002bd0008ffffffff0000440000001f0004000000013100000001340000000331323300000004352e3433440000001f0004000000013200000001340000000331323300000004352e3433430000000d53454c4543542032005a0000000549\n" +
-                    ">500000003b0073656c65637420782c24312c24322c24332066726f6d206c6f6e675f73657175656e63652832290000030000001700000014000002bd420000002600000003000000000000000300000001340000000331323300000004352e3433000044000000065000450000000900000000005300000004\n" +
-                    "<31000000043200000004540000005900047800000000000001000000140008ffffffff0000243100000000000002000000170004ffffffff0000243200000000000003000000140008ffffffff0000243300000000000004000002bd0008ffffffff0000440000001f0004000000013100000001340000000331323300000004352e3433440000001f0004000000013200000001340000000331323300000004352e3433430000000d53454c4543542032005a0000000549\n" +
-                    ">500000003e535f310073656c65637420782c24312c24322c24332066726f6d206c6f6e675f73657175656e63652832290000030000001700000014000002bd420000002900535f31000003000000000000000300000001340000000331323300000004352e3433000044000000065000450000000900000000005300000004\n" +
-                    "<31000000043200000004540000005900047800000000000001000000140008ffffffff0000243100000000000002000000170004ffffffff0000243200000000000003000000140008ffffffff0000243300000000000004000002bd0008ffffffff0000440000001f0004000000013100000001340000000331323300000004352e3433440000001f0004000000013200000001340000000331323300000004352e3433430000000d53454c4543542032005a0000000549\n" +
-                    ">420000002900535f31000003000000000000000300000001340000000331323300000004352e34330000450000000900000000005300000004\n" +
-                    "<3200000004440000001f0004000000013100000001340000000331323300000004352e3433440000001f0004000000013200000001340000000331323300000004352e3433430000000d53454c4543542032005a0000000549\n" +
-                    ">430000000953535f31005300000004\n" +
-                    "<33000000045a0000000549\n" +
-                    ">5800000004\n";
-        }
-        assertHexScriptAltCreds(
-                script
-        );
+        // new bahviour:
+        //   - the server respects the parameter types sent by the client
+        script = ">0000006e00030000757365720078797a0064617461626173650071646200636c69656e745f656e636f64696e67005554463800446174655374796c650049534f0054696d655a6f6e65004575726f70652f4c6f6e646f6e0065787472615f666c6f61745f64696769747300320000\n" +
+                "<520000000800000003\n" +
+                ">70000000076f6800\n" +
+                "<520000000800000000530000001154696d655a6f6e6500474d5400530000001d6170706c69636174696f6e5f6e616d6500517565737444420053000000187365727665725f76657273696f6e0031312e33005300000019696e74656765725f6461746574696d6573006f6e005300000019636c69656e745f656e636f64696e670055544638004b0000000c0000003fbb8b96505a0000000549\n" +
+                ">5000000022005345542065787472615f666c6f61745f646967697473203d2033000000420000000c0000000000000000450000000900000000015300000004\n" +
+                "<310000000432000000044300000008534554005a0000000549\n" +
+                ">500000003700534554206170706c69636174696f6e5f6e616d65203d2027506f737467726553514c204a4442432044726976657227000000420000000c0000000000000000450000000900000000015300000004\n" +
+                "<310000000432000000044300000008534554005a0000000549\n" +
+                ">500000003b0073656c65637420782c24312c24322c24332066726f6d206c6f6e675f73657175656e63652832290000030000001700000014000002bd420000002600000003000000000000000300000001340000000331323300000004352e3433000044000000065000450000000900000000005300000004\n" +
+                "<31000000043200000004540000005900047800000000000001000000140008ffffffff0000243100000000000002000000170004ffffffff0000243200000000000003000000140008ffffffff0000243300000000000004000002bd0008ffffffff0000440000001f0004000000013100000001340000000331323300000004352e3433440000001f0004000000013200000001340000000331323300000004352e3433430000000d53454c4543542032005a0000000549\n" +
+                ">500000003b0073656c65637420782c24312c24322c24332066726f6d206c6f6e675f73657175656e63652832290000030000001700000014000002bd420000002600000003000000000000000300000001340000000331323300000004352e3433000044000000065000450000000900000000005300000004\n" +
+                "<31000000043200000004540000005900047800000000000001000000140008ffffffff0000243100000000000002000000170004ffffffff0000243200000000000003000000140008ffffffff0000243300000000000004000002bd0008ffffffff0000440000001f0004000000013100000001340000000331323300000004352e3433440000001f0004000000013200000001340000000331323300000004352e3433430000000d53454c4543542032005a0000000549\n" +
+                ">500000003b0073656c65637420782c24312c24322c24332066726f6d206c6f6e675f73657175656e63652832290000030000001700000014000002bd420000002600000003000000000000000300000001340000000331323300000004352e3433000044000000065000450000000900000000005300000004\n" +
+                "<31000000043200000004540000005900047800000000000001000000140008ffffffff0000243100000000000002000000170004ffffffff0000243200000000000003000000140008ffffffff0000243300000000000004000002bd0008ffffffff0000440000001f0004000000013100000001340000000331323300000004352e3433440000001f0004000000013200000001340000000331323300000004352e3433430000000d53454c4543542032005a0000000549\n" +
+                ">500000003b0073656c65637420782c24312c24322c24332066726f6d206c6f6e675f73657175656e63652832290000030000001700000014000002bd420000002600000003000000000000000300000001340000000331323300000004352e3433000044000000065000450000000900000000005300000004\n" +
+                "<31000000043200000004540000005900047800000000000001000000140008ffffffff0000243100000000000002000000170004ffffffff0000243200000000000003000000140008ffffffff0000243300000000000004000002bd0008ffffffff0000440000001f0004000000013100000001340000000331323300000004352e3433440000001f0004000000013200000001340000000331323300000004352e3433430000000d53454c4543542032005a0000000549\n" +
+                ">500000003e535f310073656c65637420782c24312c24322c24332066726f6d206c6f6e675f73657175656e63652832290000030000001700000014000002bd420000002900535f31000003000000000000000300000001340000000331323300000004352e3433000044000000065000450000000900000000005300000004\n" +
+                "<31000000043200000004540000005900047800000000000001000000140008ffffffff0000243100000000000002000000170004ffffffff0000243200000000000003000000140008ffffffff0000243300000000000004000002bd0008ffffffff0000440000001f0004000000013100000001340000000331323300000004352e3433440000001f0004000000013200000001340000000331323300000004352e3433430000000d53454c4543542032005a0000000549\n" +
+                ">420000002900535f31000003000000000000000300000001340000000331323300000004352e34330000450000000900000000005300000004\n" +
+                "<3200000004440000001f0004000000013100000001340000000331323300000004352e3433440000001f0004000000013200000001340000000331323300000004352e3433430000000d53454c4543542032005a0000000549\n" +
+                ">430000000953535f31005300000004\n" +
+                "<33000000045a0000000549\n" +
+                ">5800000004\n";
+        assertHexScriptAltCreds(script);
     }
 
     @Test
@@ -2972,63 +2856,6 @@ if __name__ == "__main__":
                         ">430000000953535f31005300000004\n" +
                         "<33000000045a0000000549\n" +
                         ">5800000004\n"
-        );
-    }
-
-    @Test
-    public void testCloseMessageWithBadUtf8InStatementNameHex() throws Exception {
-        // @Ignore
-        // Test expects the server to disconnect, but modern server
-        // ignores the faulty prepared statement name and returns "ready for query"
-        Assume.assumeTrue(legacyMode);
-        skipOnWalRun(); // select only
-        assertHexScriptAltCreds(
-                ">0000006e00030000757365720078797a0064617461626173650071646200636c69656e745f656e636f64696e67005554463800446174655374796c650049534f0054696d655a6f6e65004575726f70652f4c6f6e646f6e0065787472615f666c6f61745f64696769747300320000\n" +
-                        "<520000000800000003\n" +
-                        ">70000000076f6800\n" +
-                        "<520000000800000000530000001154696d655a6f6e6500474d5400530000001d6170706c69636174696f6e5f6e616d6500517565737444420053000000187365727665725f76657273696f6e0031312e33005300000019696e74656765725f6461746574696d6573006f6e005300000019636c69656e745f656e636f64696e670055544638004b0000000c0000003fbb8b96505a0000000549\n" +
-                        ">5000000022005345542065787472615f666c6f61745f646967697473203d2033000000420000000c0000000000000000450000000900000000015300000004\n" +
-                        "<310000000432000000044300000008534554005a0000000549\n" +
-                        ">500000003700534554206170706c69636174696f6e5f6e616d65203d2027506f737467726553514c204a4442432044726976657227000000420000000c0000000000000000450000000900000000015300000004\n" +
-                        "<310000000432000000044300000008534554005a0000000549\n" +
-                        ">500000002a0073656c65637420312c322c332066726f6d206c6f6e675f73657175656e6365283129000000420000000c000000000000000044000000065000450000000900000000005300000004\n" +
-                        "<31000000043200000004540000004200033100000000000001000000170004ffffffff00003200000000000002000000170004ffffffff00003300000000000003000000170004ffffffff000044000000150003000000013100000001320000000133430000000d53454c4543542031005a0000000549\n" +
-                        ">500000002d535f310073656c65637420312c322c332066726f6d206c6f6e675f73657175656e6365283129000000420000000f00535f310000000000000044000000065000450000000900000000005300000004\n" +
-                        "<31000000043200000004540000004200033100000000000001000000170004ffffffff00003200000000000002000000170004ffffffff00003300000000000003000000170004ffffffff000044000000150003000000013100000001320000000133430000000d53454c4543542031005a0000000549\n" +
-                        ">420000000f00535f3100000000000000450000000900000000005300000004\n" +
-                        "<320000000444000000150003000000013100000001320000000133430000000d53454c4543542031005a0000000549\n" +
-                        ">420000000f00535f3100000000000000450000000900000000005300000004\n" +
-                        "<320000000444000000150003000000013100000001320000000133430000000d53454c4543542031005a0000000549\n" +
-                        ">430000000953535fac005300000004\n" +
-                        "<!!"
-        );
-    }
-
-    @Test
-    public void testCloseMessageWithInvalidTypeHex() throws Exception {
-        // Test expects the server to disconnect, but modern server
-        // responds with error "invalid type for close message [type=81]", followed by "Ready for query"
-        Assume.assumeTrue(legacyMode);
-        skipOnWalRun(); // select only
-        assertHexScriptAltCreds(
-                ">0000006e00030000757365720078797a0064617461626173650071646200636c69656e745f656e636f64696e67005554463800446174655374796c650049534f0054696d655a6f6e65004575726f70652f4c6f6e646f6e0065787472615f666c6f61745f64696769747300320000\n" +
-                        "<520000000800000003\n" +
-                        ">70000000076f6800\n" +
-                        "<520000000800000000530000001154696d655a6f6e6500474d5400530000001d6170706c69636174696f6e5f6e616d6500517565737444420053000000187365727665725f76657273696f6e0031312e33005300000019696e74656765725f6461746574696d6573006f6e005300000019636c69656e745f656e636f64696e670055544638004b0000000c0000003fbb8b96505a0000000549\n" +
-                        ">5000000022005345542065787472615f666c6f61745f646967697473203d2033000000420000000c0000000000000000450000000900000000015300000004\n" +
-                        "<310000000432000000044300000008534554005a0000000549\n" +
-                        ">500000003700534554206170706c69636174696f6e5f6e616d65203d2027506f737467726553514c204a4442432044726976657227000000420000000c0000000000000000450000000900000000015300000004\n" +
-                        "<310000000432000000044300000008534554005a0000000549\n" +
-                        ">500000002a0073656c65637420312c322c332066726f6d206c6f6e675f73657175656e6365283129000000420000000c000000000000000044000000065000450000000900000000005300000004\n" +
-                        "<31000000043200000004540000004200033100000000000001000000170004ffffffff00003200000000000002000000170004ffffffff00003300000000000003000000170004ffffffff000044000000150003000000013100000001320000000133430000000d53454c4543542031005a0000000549\n" +
-                        ">500000002d535f310073656c65637420312c322c332066726f6d206c6f6e675f73657175656e6365283129000000420000000f00535f310000000000000044000000065000450000000900000000005300000004\n" +
-                        "<31000000043200000004540000004200033100000000000001000000170004ffffffff00003200000000000002000000170004ffffffff00003300000000000003000000170004ffffffff000044000000150003000000013100000001320000000133430000000d53454c4543542031005a0000000549\n" +
-                        ">420000000f00535f3100000000000000450000000900000000005300000004\n" +
-                        "<320000000444000000150003000000013100000001320000000133430000000d53454c4543542031005a0000000549\n" +
-                        ">420000000f00535f3100000000000000450000000900000000005300000004\n" +
-                        "<320000000444000000150003000000013100000001320000000133430000000d53454c4543542031005a0000000549\n" +
-                        ">430000000951535f31005300000004\n" +
-                        "<!!"
         );
     }
 
@@ -3325,7 +3152,6 @@ if __name__ == "__main__":
 
     @Test
     public void testErrnoInErrorMessage() throws Exception {
-        Assume.assumeFalse(legacyMode); // only modern server prints errno
         skipOnWalRun(); // non-partitioned table
         ff = new TestFilesFacadeImpl() {
             @Override
@@ -4362,352 +4188,233 @@ if __name__ == "__main__":
         // is specified as OID 700 on the Row Description, but is sent as 0.46218354.
         // The modern code sends FLOAT as FLOAT, e.g. 0.462 instead.
         final String script;
-        if (legacyMode) {
-            script = ">0000000804d2162f\n" +
-                    "<4e\n" +
-                    ">0000007000030000757365720061646d696e0064617461626173650071646200636c69656e745f656e636f64696e670055544638004461746553\n" +
-                    ">74796c65004953\n" +
-                    ">4f00\n" +
-                    ">54696d655a6f6e6500\n" +
-                    ">4575\n" +
-                    ">726f70652f4c6f\n" +
-                    ">6e646f6e0065787472615f666c6f\n" +
-                    ">61745f64\n" +
-                    ">696769\n" +
-                    ">747300320000\n" +
-                    "<520000000800000003\n" +
-                    ">7000\n" +
-                    ">00000a\n" +
-                    ">717565\n" +
-                    ">737400\n" +
-                    "<520000000800000000530000001154696d655a6f6e6500474d5400530000001d6170706c69636174696f6e5f6e616d6500517565737444420053000000187365727665725f76657273696f6e0031312e33005300000019696e74656765725f6461746574696d6573006f6e005300000019636c69656e745f656e636f64696e670055544638004b0000000c0000003fbb8b96505a0000000549\n" +
-                    ">50\n" +
-                    ">00000022005345542065787472615f666c6f61745f64696769747320\n" +
-                    ">3d2033\n" +
-                    ">00\n" +
-                    ">0000420000000c0000000000000000450000000900000000015300000004\n" +
-                    "<310000000432000000044300000008534554005a0000000549\n" +
-                    ">500000003700\n" +
-                    ">5345\n" +
-                    ">54206170706c69636174696f6e5f\n" +
-                    ">6e616d65203d2027506f737467726553514c204a4442432044726976657227\n" +
-                    ">000000\n" +
-                    ">420000000c0000000000000000450000000900000000015300000004\n" +
-                    "<310000000432000000044300000008534554005a0000000549\n" +
-                    ">5000\n" +
-                    ">0001940073656c656374\n" +
-                    ">20726e645f7374722834\n" +
-                    ">2c342c\n" +
-                    ">342920732c\n" +
-                    ">20\n" +
-                    ">726e\n" +
-                    ">645f696e7428\n" +
-                    ">302c20\n" +
-                    ">32\n" +
-                    ">35362c20342920\n" +
-                    ">692c\n" +
-                    ">20\n" +
-                    ">726e645f646f75\n" +
-                    ">626c65\n" +
-                    ">28342920642c207469\n" +
-                    ">6d6573\n" +
-                    ">7461\n" +
-                    ">6d705f73657175656e636528302c3130\n" +
-                    ">30\n" +
-                    ">30302920742c20726e645f666c6f617428\n" +
-                    ">342920\n" +
-                    ">66\n" +
-                    ">2c20726e\n" +
-                    ">645f73\n" +
-                    ">686f72742829205f73\n" +
-                    ">686f7274\n" +
-                    ">2c2072\n" +
-                    ">6e645f\n" +
-                    ">6c6f6e6728302c203130303030\n" +
-                    ">3030302c2035\n" +
-                    ">29206c2c20726e64\n" +
-                    ">5f74696d657374616d70\n" +
-                    ">28746f\n" +
-                    ">5f\n" +
-                    ">74696d65737461\n" +
-                    ">6d70282732303135272c277979797927292c746f5f74696d657374616d\n" +
-                    ">7028273230313627\n" +
-                    ">2c\n" +
-                    ">277979797927292c3229207473322c2072\n" +
-                    ">6e645f62\n" +
-                    ">79\n" +
-                    ">746528\n" +
-                    ">302c31323729206262\n" +
-                    ">2c20\n" +
-                    ">726e\n" +
-                    ">64\n" +
-                    ">5f626f6f6c\n" +
-                    ">6561\n" +
-                    ">6e282920\n" +
-                    ">622c20726e64\n" +
-                    ">5f73\n" +
-                    ">796d626f6c2834\n" +
-                    ">2c342c342c32292c\n" +
-                    ">20\n" +
-                    ">726e645f646174\n" +
-                    ">6528\n" +
-                    ">74\n" +
-                    ">6f5f646174652827323031\n" +
-                    ">35272c202779\n" +
-                    ">7979\n" +
-                    ">7927292c20\n" +
-                    ">746f5f64\n" +
-                    ">61746528\n" +
-                    ">27\n" +
-                    ">32\n" +
-                    ">30\n" +
-                    ">3136272c2027797979\n" +
-                    ">792729\n" +
-                    ">2c\n" +
-                    ">20\n" +
-                    ">3229\n" +
-                    ">2c726e645f62\n" +
-                    ">69\n" +
-                    ">6e2831302c32\n" +
-                    ">302c322920\n" +
-                    ">66726f6d\n" +
-                    ">206c6f\n" +
-                    ">6e675f7365\n" +
-                    ">7175656e6365283530\n" +
-                    ">29\n" +
-                    ">0000004200\n" +
-                    ">00000c000000000000000044000000065000450000000900000000005300000004\n" +
-                    "<310000000432000000045400000128000d730000000000000100000413ffffffffffff00006900000000000002000000170004ffffffff00006400000000000003000002bd0008ffffffff000074000000000000040000045a0008ffffffff00006600000000000005000002bc0004ffffffff00005f73686f727400000000000006000000150002ffffffff00006c00000000000007000000140008ffffffff0000747332000000000000080000045a0008ffffffff0000626200000000000009000000150001ffffffff0000620000000000000a000000100001ffffffff0000726e645f73796d626f6c0000000000000b00000413ffffffffffff0000726e645f646174650000000000000c0000045a0008ffffffff0000726e645f62696e0000000000000d00000011ffffffffffff000144000000ab000dffffffff00000002353700000012302e363235343032313534323431323031380000001a313937302d30312d30312030303a30303a30302e3030303030300000000a302e3436323138333534000000052d313539330000000733343235323332ffffffff000000033132310000000166000000045045484e00000017323031352d30332d31372030343a32353a35322e3736350000000e19c49594365349b4597e3b08a11e44000000cc000d00000004585953420000000331343200000012302e353739333436363332363836323231310000001a313937302d30312d30312030303a30303a30302e30313030303000000009302e3936383734323300000005323030383800000007313531373439300000001a323031352d30312d31372032303a34313a31392e343830363835000000033130300000000174000000045045484e00000017323031352d30362d32302030313a31303a35382e35393900000011795f8b812b934d1a8e78b5b91153d0fb6444000000c7000d000000044f5a5a560000000332313900000013302e31363338313337343737333734383531340000001a313937302d30312d30312030303a30303a30302e3032303030300000000a302e3635393033343133000000062d313233303300000007393438393530380000001a323031352d30382d31332031373a31303a31392e37353235323100000001360000000166ffffffff00000017323031352d30352d32302030313a34383a33372e3431380000000f2b4d5ff64690c3b3598ee5612f640e44000000b6000d000000044f4c595800000002333000000012302e373133333931303237313535353834330000001a313937302d30312d30312030303a30303a30302e3033303030300000000a302e3635353133333535000000043636313000000007363530343432380000001a323031352d30382d30382030303a34323a32342e353435363339000000033132330000000166ffffffff00000017323031352d30312d30332031333a35333a30332e313635ffffffff44000000af000d000000045449514200000002343200000012302e363830363837333133343632363431380000001a313937302d30312d30312030303a30303a30302e30343030303000000008302e363235393636000000052d3136303500000007383831343038360000001a323031352d30372d32382031353a30383a35332e34363234393500000002323800000001740000000443505357ffffffff0000000e3ba6dc3b7d2be392fe6938e1779a44000000b4000d000000044c544f560000000331333700000012302e373633323631353030343332343530330000001a313937302d30312d30312030303a30303a30302e3035303030300000000a302e38383136393034340000000439303534ffffffff0000001a323031352d30342d32302030353a30393a30332e353830353734000000033130360000000166000000045045484e00000017323031352d30312d30392030363a35373a31372e353132ffffffff44000000a0000d000000045a494d4e00000003313235ffffffff0000001a313937302d30312d30312030303a30303a30302e303630303030ffffffff00000005313135323400000007383333353236310000001a323031352d31302d32362030323a31303a35302e363838333934000000033131310000000174000000045045484e00000017323031352d30382d32312031353a34363a33322e363234ffffffff44000000a5000d000000044f504a4f0000000331363800000013302e31303435393335323331323333313138330000001a313937302d30312d30312030303a30303a30302e30373030303000000009302e35333436303139000000052d3539323000000007373038303730340000001a323031352d30372d31312030393a31353a33382e3334323731370000000331303300000001660000000456544a57ffffffffffffffff44000000bb000d00000004474c554f0000000331343500000012302e353339313632363632313739343637330000001a313937302d30312d30312030303a30303a30302e3038303030300000000a302e373636383134363500000005313432343200000007323439393932320000001a323031352d31312d30322030393a30313a33312e3331323830340000000238340000000166000000045045484e00000017323031352d31312d31342031373a33373a33362e303433ffffffff44000000c3000d000000045a5651450000000331303300000012302e363732393430353539303737333633380000001a313937302d30312d30312030303a30303a30302e303930303030ffffffff00000005313337323700000007373837353834360000001a323031352d31322d31322031333a31363a32362e3133343536320000000232320000000174000000045045484e00000017323031352d30312d32302030343a35303a33342e30393800000012143380c9eba3677a1a79e435e43adc5c65ff44000000a7000d000000044c4947590000000331393900000012302e323833363334373133393438313436390000001a313937302d30312d30312030303a30303a30302e313030303030ffffffff00000005333034323600000007333231353536320000001a323031352d30382d32312031343a35353a30372e30353537323200000002313100000001660000000456544a57ffffffff0000000dff703ac78ab314cd470b0c391244000000ac000d000000044d514e5400000002343300000012302e353835393333323338383539393633380000001a313937302d30312d30312030303a30303a30302e3131303030300000000a302e3333353034313436000000053237303139ffffffffffffffff0000000232370000000174000000045045484e00000017323031352d30372d31322031323a35393a34372e3636350000001326fb2e42faf56e8f80e354b807b13257ff9aef44000000cd000d00000004575743430000000332313300000012302e373636353032393931343337363935320000001a313937302d30312d30312030303a30303a30302e3132303030300000000a302e353739363732313600000005313336343000000007343132313932330000001a323031352d30382d30362030323a32373a33302e3436393736320000000237330000000166000000045045484e00000017323031352d30342d33302030383a31383a31302e3435330000001271a7d5af11963708dd98ef54882aa2ade7d444000000b2000d00000004564647500000000331323000000012302e383430323936343730383132393534360000001a313937302d30312d30312030303a30303a30302e31333030303000000008302e373733323233000000043732323300000007373234313432330000001a323031352d31322d31382030373a33323a31382e34353630323500000002343300000001660000000456544a57ffffffff00000011244e44a80dfe27ec53135db215e7b8356744000000bc000d00000004524d44470000000331333400000013302e31313034373331353231343739333639360000001a313937302d30312d30312030303a30303a30302e3134303030300000000a302e303433323132383900000005323132323700000007373135353730380000001a323031352d30372d30332030343a31323a34352e3737343238310000000234320000000174000000044350535700000017323031352d30322d32342031323a31303a34332e313939ffffffff44000000aa000d0000000457464f5100000003323535ffffffff0000001a313937302d30312d30312030303a30303a30302e3135303030300000000a302e313136323432343700000005333135363900000007363638383237370000001a323031352d30352d31392030333a33303a34352e373739393939000000033132360000000174000000045045484e00000017323031352d31322d30392030393a35373a31372e303738ffffffff440000009d000d000000044d58444b00000002353600000012302e393939373739373233343033313638380000001a313937302d30312d30312030303a30303a30302e3136303030300000000a302e3532333438393234000000062d33323337320000000736383834313332ffffffff0000000235380000000166ffffffff00000017323031352d30312d32302030363a31383a31382e353833ffffffff44000000bf000d00000004584d4b4a0000000331333900000012302e383430353831353439333536373431370000001a313937302d30312d30312030303a30303a30302e31373030303000000009302e33303538303038000000053235383536ffffffff0000001a323031352d30352d31382030333a35303a32322e373331343337000000013200000001740000000456544a5700000017323031352d30362d32352031303a34353a30312e3031340000000d007cfb0119caf2bf845a6f383544000000b4000d0000000456494844ffffffffffffffff0000001a313937302d30312d30312030303a30303a30302e3138303030300000000a302e353530313133323600000005323232383000000007393130393834320000001a323031352d30312d32352031333a35313a33382e3237303538330000000239340000000166000000044350535700000017323031352d31302d32372030323a35323a31392e3933350000000e2d16f389a38364ded6fdc45bc4e944000000c1000d0000000457504e58ffffffff00000012302e393436393730303831333932363930370000001a313937302d30312d30312030303a30303a30302e31393030303000000009302e34313439363631000000062d3137393333000000063637343236310000001a323031352d30332d30342031353a34333a31352e3231333638360000000234330000000174000000044859525800000017323031352d31322d31382032313a32383a32352e3332350000000ab34c0e8ff10cc560b7d144000000c3000d0000000459504f5600000002333600000012302e363734313234383434383732383832340000001a313937302d30312d30312030303a30303a30302e3230303030300000000b302e303330393937333936000000052d3538383800000007313337353432330000001a323031352d31322d31302032303a35303a33352e38363636313400000001330000000174ffffffff00000017323031352d30372d32332032303a31373a30342e3233360000000dd4abbe30fa8dac3d98a0ad9a5d44000000cb000d000000044e55484effffffff00000012302e363934303931373932353134383333320000001a313937302d30312d30312030303a30303a30302e3231303030300000000a302e3333393234383138000000062d323532323600000007333532343734380000001a323031352d30352d30372030343a30373a31382e31353239363800000002333900000001740000000456544a5700000017323031352d30342d30342031353a32333a33342e31333000000012b8bef8a146872892a39be3cbc2648ab035d844000000a0000d00000004424f53450000000332343000000013302e30363030313832373732313535363031390000001a313937302d30312d30312030303a30303a30302e32323030303000000009302e3337383739323700000005323339303400000007393036393333390000001a323031352d30332d32312030333a34323a34322e3634333138360000000238340000000174ffffffffffffffffffffffff44000000c9000d00000004494e4b470000000331323400000012302e383631353834313632373730323735330000001a313937302d30312d30312030303a30303a30302e32333030303000000009302e34303431303136000000062d333033383300000007373233333534320000001a323031352d30372d32312031363a34323a34372e3031323134380000000239390000000166ffffffff00000017323031352d30382d32372031373a32353a33352e3330380000001287fc9283fc88f3322770c801b0dcc93a5b7e44000000b1000d000000044655584300000002353200000012302e373433303130313939343531313531370000001a313937302d30312d30312030303a30303a30302e323430303030ffffffff000000062d313437323900000007313034323036340000001a323031352d30382d32312030323a31303a35382e3934393637340000000232380000000174000000044350535700000017323031352d30382d32392032303a31353a35312e383335ffffffff44000000c2000d00000004554e595100000002373100000011302e3434323039353431303238313933380000001a313937302d30312d30312030303a30303a30302e3235303030300000000a302e3533393435363235000000062d3232363131ffffffff0000001a323031352d31322d32332031383a34313a34322e3331393835390000000239380000000174000000045045484e00000017323031352d30312d32362030303a35353a35302e3230320000000f28ed9799d877333fb267da984747bf44000000b1000d000000044b424d51ffffffff00000013302e32383031393231383832353035313339350000001a313937302d30312d30312030303a30303a30302e323630303030ffffffff000000053132323430ffffffff0000001a323031352d30382d31362030313a30323a35352e3736363632320000000232310000000166ffffffff00000017323031352d30352d31392030303a34373a31382e3639380000000d6ade4604d381e7a21622353b1c4400000096000d000000044a534f4c00000003323433ffffffff0000001a313937302d30312d30312030303a30303a30302e3237303030300000000a302e3036383230313636000000062d3137343638ffffffffffffffff0000000232300000000174ffffffff00000017323031352d30362d31392031303a33383a35342e343833000000113de02d0486e7ca29980769ca5bd6cf09694400000084000d00000004484e535300000003313530ffffffff0000001a313937302d30312d30312030303a30303a30302e3238303030300000000a302e31343735363335320000000531343834310000000735393932343433ffffffff0000000232350000000166000000045045484effffffff0000000c14d6fcee032281b806c406af44000000c3000d00000004505a50420000000331303100000014302e3036313634363731373738363135383034350000001a313937302d30312d30312030303a30303a30302e323930303030ffffffff00000005313232333700000007393837383137390000001a323031352d30392d30332032323a31333a31382e38353234363500000002373900000001660000000456544a5700000017323031352d31322d31372031353a31323a35342e3935380000001012613a9aad982e7552ad62878845b99d44000000c8000d000000044f594e4e00000002323500000012302e333339333530393531343030303234370000001a313937302d30312d30312030303a30303a30302e3330303030300000000a302e363238313235323500000005323234313200000007343733363337380000001a323031352d31302d31302031323a31393a34322e353238323234000000033130360000000174000000044350535700000017323031352d30372d30312030303a32333a34392e3738390000000d54133fffb67ecd0427669489db4400000083000dffffffff0000000331313700000012302e353633383430343737353636333136310000001a313937302d30312d30312030303a30303a30302e333130303030ffffffff000000052d353630340000000736333533303138ffffffff0000000238340000000166ffffffffffffffff0000000b2bad2507db6244336e008e440000009d000d00000004485652490000000332333300000013302e32323430373636353739303730353737370000001a313937302d30312d30312030303a30303a30302e33323030303000000009302e343234363635310000000531303436390000000731373135323133ffffffff0000000238360000000166ffffffff00000017323031352d30322d30322030353a34383a31372e333733ffffffff44000000bb000d000000044f59544f00000002393600000012302e373430373538313631363931363336340000001a313937302d30312d30312030303a30303a30302e3333303030300000000a302e3532373737363636000000062d313232333900000007333439393632300000001a323031352d30322d30372032323a33353a30332e3231323236380000000231370000000166000000045045484e00000017323031352d30332d32392031323a35353a31312e363832ffffffff44000000a5000d000000044c46435900000002363300000012302e373231373331353732393739303732320000001a313937302d30312d30312030303a30303a30302e333430303030ffffffff0000000532333334340000000739353233393832ffffffff000000033132330000000166000000044350535700000017323031352d30352d31382030343a33353a32372e3232380000000e05e5c04eccd6e37b34cd1535bba444000000c4000d0000000447484c580000000331343800000012302e333035373933373730343936343237320000001a313937302d30312d30312030303a30303a30302e33353030303000000008302e363335353539000000062d333134353700000007323332323333370000001a323031352d31302d32322031323a30363a30352e3534343730310000000239310000000174000000044859525800000017323031352d30352d32312030393a33333a31382e3135380000000a571d91723004b702cb0344000000a9000d000000045954535a00000003313233ffffffff0000001a313937302d30312d30312030303a30303a30302e3336303030300000000a302e353139313838343600000005323235333400000007343434363233360000001a323031352d30372d32372030373a32333a33372e3233333731310000000235330000000166000000044350535700000017323031352d30312d31332030343a33373a31302e303336ffffffff44000000a8000d0000000453574c5500000003323531ffffffff0000001a313937302d30312d30312030303a30303a30302e3337303030300000000a302e3137393034373538000000043737333400000007343038323437350000001a323031352d31302d32312031383a32343a33342e3430303334350000000236390000000166000000045045484e00000017323031352d30342d30312031343a33333a34322e303035ffffffff44000000b5000d0000000454514a4c00000003323435ffffffff0000001a313937302d30312d30312030303a30303a30302e33383030303000000009302e383634353533360000000439353136000000063932393334300000001a323031352d30352d32382030343a31383a31382e36343035363700000002363900000001660000000456544a5700000017323031352d30362d31322032303a31323a32382e3838310000000f6c3e51d7ebb10771321faf404e8c4744000000a3000d000000045245494a000000023934ffffffff0000001a313937302d30312d30312030303a30303a30302e3339303030300000000a302e3133303237373939000000062d3239393234ffffffff0000001a323031352d30332d32302032323a31343a34362e323034373138000000033131330000000174000000044859525800000017323031352d31322d31392031333a35383a34312e383139ffffffff44000000c7000d000000044844485100000002393400000012302e373233343138313737333430373533360000001a313937302d30312d30312030303a30303a30302e3430303030300000000a302e3732393733333335000000053139393730000000063635343133310000001a323031352d30312d31302032323a35363a30382e3438303435300000000238340000000174ffffffff00000017323031352d30332d30352031373a31343a34382e323735000000124f566b65a45338e9cdc1a7ee8675ada52d4944000000bd000d00000004554d455500000002343000000014302e3030383434343033333233303538303733390000001a313937302d30312d30312030303a30303a30302e3431303030300000000a302e3830353237323736000000062d313136323300000007343539393836320000001a323031352d31312d32302030343a30323a34342e3333353934370000000237360000000166000000045045484e00000017323031352d30352d31372031373a33333a32302e393232ffffffff44000000b2000d00000004594a494800000003313834ffffffff0000001a313937302d30312d30312030303a30303a30302e3432303030300000000a302e333832363936363300000005313736313400000007333130313637310000001a323031352d30312d32382031323a30353a34362e363833303031000000033130350000000174ffffffff00000017323031352d31322d30372031393a32343a33362e3833380000000cec69cd73bb9bc595db6191ce44000000a7000d000000044359584700000002323700000012302e323931373739363035333034353734370000001a313937302d30312d30312030303a30303a30302e34333030303000000009302e39353239313736000000043339343400000006323439313635ffffffff0000000236370000000174ffffffff00000017323031352d30332d30322030383a31393a34342e3536360000000e0148153e0c7f3f8fe4b5ab34212944000000b8000d000000044d5254470000000331343300000013302e30323633323533313336313439393131330000001a313937302d30312d30312030303a30303a30302e34343030303000000009302e39343235363538000000062d323733323000000007313636373834320000001a323031352d30312d32342031393a35363a31352e3937333130390000000231310000000166ffffffff00000017323031352d30312d32342030373a31353a30322e373732ffffffff44000000c7000d00000004444f4e500000000332343600000011302e3635343232363234383734303434370000001a313937302d30312d30312030303a30303a30302e34353030303000000009302e3535353736313700000005323734373700000007343136303031380000001a323031352d31322d31342030333a34303a30352e3931313833390000000232300000000174000000045045484e00000017323031352d31302d32392031343a33353a31302e3136370000000e079201f56aa131cdcbc2a2b48e9944000000ca000d00000004495158530000000332333200000013302e32333037353730303231383033383835330000001a313937302d30312d30312030303a30303a30302e3436303030300000000b302e303438353430323335000000062d313831313300000007343030353232380000001a323031352d30362d31312031333a30303a30372e32343831383800000001380000000174000000044350535700000017323031352d30382d31362031313a30393a32342e3331310000000dfa1f9224b1b8676508b7f8410044000000b4000dffffffff00000003313738ffffffff0000001a313937302d30312d30312030303a30303a30302e34373030303000000008302e393033303532000000062d313436323600000007323933343537300000001a323031352d30342d30342030383a35313a35342e3036383135340000000238380000000174ffffffff00000017323031352d30372d30312030343a33323a32332e30383300000014843625632b6361431c477db646babb98ca08bea444000000b5000d000000044855575a00000002393400000011302e3131303430313337343937393631330000001a313937302d30312d30312030303a30303a30302e3438303030300000000a302e3432303338393737000000052d3337333600000007353638373531340000001a323031352d30312d30322031373a31383a30352e3632373633330000000237340000000166ffffffff00000017323031352d30332d32392030363a33393a31312e363432ffffffff44000000b1000d000000045352454400000002363600000013302e31313237343636373134303931353932380000001a313937302d30312d30312030303a30303a30302e3439303030300000000b302e303539383639383236000000062d313035343300000007333636393337370000001a323031352d31302d32322030323a35333a30322e3338313335310000000237370000000174000000045045484effffffff0000000b7c3fd6883a93ef24a5e2bc430000000e53454c454354203530005a0000000549\n" +
-                    ">5800000004\n";
-        } else {
-            script = ">0000000804d2162f\n" +
-                    "<4e\n" +
-                    ">000000700003000075736572\n" +
-                    ">0061646d696e0064617461\n" +
-                    ">626173650071\n" +
-                    ">646200636c\n" +
-                    ">69656e74\n" +
-                    ">5f656e636f\n" +
-                    ">64696e6700\n" +
-                    ">5554463800446174\n" +
-                    ">655374796c650049\n" +
-                    ">534f0054696d655a\n" +
-                    ">6f6e65004575\n" +
-                    ">726f70652f\n" +
-                    ">4c6f6e64\n" +
-                    ">6f6e006578\n" +
-                    ">7472615f666c6f\n" +
-                    ">61745f64\n" +
-                    ">696769747300320000\n" +
-                    "<520000000800000003\n" +
-                    ">700000\n" +
-                    ">000a7175\n" +
-                    ">65737400\n" +
-                    "<520000000800000000530000001154696d655a6f6e6500474d5400530000001d6170706c69636174696f6e5f6e616d6500517565737444420053000000187365727665725f76657273696f6e0031312e33005300000019696e74656765725f6461746574696d6573006f6e005300000019636c69656e745f656e636f64696e670055544638004b0000000c0000003fbb8b96505a0000000549\n" +
-                    ">50000000220053455420\n" +
-                    ">65787472615f666c6f61745f64\n" +
-                    ">69\n" +
-                    ">67697473203d2033000000420000\n" +
-                    ">000c0000000000000000450000000900000000015300000004\n" +
-                    "<310000000432000000044300000008534554005a0000000549\n" +
-                    ">500000003700\n" +
-                    ">53\n" +
-                    ">45542061\n" +
-                    ">70\n" +
-                    ">70\n" +
-                    ">6c6963\n" +
-                    ">61\n" +
-                    ">74\n" +
-                    ">696f6e5f6e616d6520\n" +
-                    ">3d\n" +
-                    ">2027506f737467726553514c204a44\n" +
-                    ">424320\n" +
-                    ">44\n" +
-                    ">7269766572\n" +
-                    ">2700000042\n" +
-                    ">0000000c0000000000000000450000000900000000015300000004\n" +
-                    "<310000000432000000044300000008534554005a0000000549\n" +
-                    ">50\n" +
-                    ">000001\n" +
-                    ">94\n" +
-                    ">0073656c\n" +
-                    ">65\n" +
-                    ">637420726e\n" +
-                    ">64\n" +
-                    ">5f737472\n" +
-                    ">28\n" +
-                    ">342c342c34292073\n" +
-                    ">2c20\n" +
-                    ">726e64\n" +
-                    ">5f\n" +
-                    ">69\n" +
-                    ">6e7428302c203235362c20342920692c20726e\n" +
-                    ">645f64\n" +
-                    ">6f75626c65\n" +
-                    ">28\n" +
-                    ">34\n" +
-                    ">2920642c\n" +
-                    ">20\n" +
-                    ">74\n" +
-                    ">696d6573\n" +
-                    ">74\n" +
-                    ">616d705f\n" +
-                    ">73\n" +
-                    ">65717565\n" +
-                    ">6e63\n" +
-                    ">6528302c\n" +
-                    ">31\n" +
-                    ">30\n" +
-                    ">303030\n" +
-                    ">29\n" +
-                    ">20\n" +
-                    ">742c20726e\n" +
-                    ">64\n" +
-                    ">5f\n" +
-                    ">666c6f61\n" +
-                    ">74\n" +
-                    ">28\n" +
-                    ">34292066\n" +
-                    ">2c20726e\n" +
-                    ">64\n" +
-                    ">5f73686f\n" +
-                    ">72\n" +
-                    ">74282920\n" +
-                    ">5f\n" +
-                    ">73\n" +
-                    ">686f72742c\n" +
-                    ">20\n" +
-                    ">726e645f\n" +
-                    ">6c\n" +
-                    ">6f6e67\n" +
-                    ">2830\n" +
-                    ">2c20313030\n" +
-                    ">30\n" +
-                    ">30\n" +
-                    ">303030\n" +
-                    ">2c20\n" +
-                    ">3529206c\n" +
-                    ">2c\n" +
-                    ">20\n" +
-                    ">726e645f\n" +
-                    ">74\n" +
-                    ">696d6573\n" +
-                    ">74\n" +
-                    ">61\n" +
-                    ">6d7028\n" +
-                    ">74\n" +
-                    ">6f\n" +
-                    ">5f74696d\n" +
-                    ">65\n" +
-                    ">7374616d\n" +
-                    ">70\n" +
-                    ">28\n" +
-                    ">27323031\n" +
-                    ">35\n" +
-                    ">272c27797979\n" +
-                    ">79\n" +
-                    ">27292c746f\n" +
-                    ">5f\n" +
-                    ">74696d6573\n" +
-                    ">74\n" +
-                    ">616d702827\n" +
-                    ">32\n" +
-                    ">303136\n" +
-                    ">272c\n" +
-                    ">277979\n" +
-                    ">79\n" +
-                    ">79\n" +
-                    ">27292c\n" +
-                    ">32\n" +
-                    ">29\n" +
-                    ">207473\n" +
-                    ">32\n" +
-                    ">2c\n" +
-                    ">20726e\n" +
-                    ">64\n" +
-                    ">5f\n" +
-                    ">627974\n" +
-                    ">6528\n" +
-                    ">302c31\n" +
-                    ">32\n" +
-                    ">37\n" +
-                    ">29206262\n" +
-                    ">2c\n" +
-                    ">20726e\n" +
-                    ">64\n" +
-                    ">5f626f\n" +
-                    ">6f\n" +
-                    ">6c\n" +
-                    ">65616e\n" +
-                    ">28\n" +
-                    ">29\n" +
-                    ">20622c\n" +
-                    ">20\n" +
-                    ">72\n" +
-                    ">6e645f\n" +
-                    ">73\n" +
-                    ">79\n" +
-                    ">6d626f\n" +
-                    ">6c\n" +
-                    ">28342c34\n" +
-                    ">2c\n" +
-                    ">34\n" +
-                    ">2c3229\n" +
-                    ">2c\n" +
-                    ">20\n" +
-                    ">726e64\n" +
-                    ">5f\n" +
-                    ">646174\n" +
-                    ">65\n" +
-                    ">28\n" +
-                    ">746f5f6461\n" +
-                    ">74\n" +
-                    ">652827\n" +
-                    ">32\n" +
-                    ">303135\n" +
-                    ">27\n" +
-                    ">2c\n" +
-                    ">20\n" +
-                    ">277979\n" +
-                    ">7979\n" +
-                    ">27292c20\n" +
-                    ">74\n" +
-                    ">6f\n" +
-                    ">5f6461\n" +
-                    ">74\n" +
-                    ">652827\n" +
-                    ">32\n" +
-                    ">30\n" +
-                    ">3136272c\n" +
-                    ">2027\n" +
-                    ">797979\n" +
-                    ">79\n" +
-                    ">27292c20\n" +
-                    ">32\n" +
-                    ">29\n" +
-                    ">2c726e\n" +
-                    ">64\n" +
-                    ">5f62696e\n" +
-                    ">28\n" +
-                    ">31302c32\n" +
-                    ">30\n" +
-                    ">2c3229\n" +
-                    ">20\n" +
-                    ">66726f\n" +
-                    ">6d\n" +
-                    ">20\n" +
-                    ">6c6f6e\n" +
-                    ">67\n" +
-                    ">5f\n" +
-                    ">736571\n" +
-                    ">75\n" +
-                    ">656e63\n" +
-                    ">65\n" +
-                    ">28\n" +
-                    ">35302900\n" +
-                    ">00\n" +
-                    ">004200\n";
-        }
+        script = ">0000000804d2162f\n" +
+                "<4e\n" +
+                ">000000700003000075736572\n" +
+                ">0061646d696e0064617461\n" +
+                ">626173650071\n" +
+                ">646200636c\n" +
+                ">69656e74\n" +
+                ">5f656e636f\n" +
+                ">64696e6700\n" +
+                ">5554463800446174\n" +
+                ">655374796c650049\n" +
+                ">534f0054696d655a\n" +
+                ">6f6e65004575\n" +
+                ">726f70652f\n" +
+                ">4c6f6e64\n" +
+                ">6f6e006578\n" +
+                ">7472615f666c6f\n" +
+                ">61745f64\n" +
+                ">696769747300320000\n" +
+                "<520000000800000003\n" +
+                ">700000\n" +
+                ">000a7175\n" +
+                ">65737400\n" +
+                "<520000000800000000530000001154696d655a6f6e6500474d5400530000001d6170706c69636174696f6e5f6e616d6500517565737444420053000000187365727665725f76657273696f6e0031312e33005300000019696e74656765725f6461746574696d6573006f6e005300000019636c69656e745f656e636f64696e670055544638004b0000000c0000003fbb8b96505a0000000549\n" +
+                ">50000000220053455420\n" +
+                ">65787472615f666c6f61745f64\n" +
+                ">69\n" +
+                ">67697473203d2033000000420000\n" +
+                ">000c0000000000000000450000000900000000015300000004\n" +
+                "<310000000432000000044300000008534554005a0000000549\n" +
+                ">500000003700\n" +
+                ">53\n" +
+                ">45542061\n" +
+                ">70\n" +
+                ">70\n" +
+                ">6c6963\n" +
+                ">61\n" +
+                ">74\n" +
+                ">696f6e5f6e616d6520\n" +
+                ">3d\n" +
+                ">2027506f737467726553514c204a44\n" +
+                ">424320\n" +
+                ">44\n" +
+                ">7269766572\n" +
+                ">2700000042\n" +
+                ">0000000c0000000000000000450000000900000000015300000004\n" +
+                "<310000000432000000044300000008534554005a0000000549\n" +
+                ">50\n" +
+                ">000001\n" +
+                ">94\n" +
+                ">0073656c\n" +
+                ">65\n" +
+                ">637420726e\n" +
+                ">64\n" +
+                ">5f737472\n" +
+                ">28\n" +
+                ">342c342c34292073\n" +
+                ">2c20\n" +
+                ">726e64\n" +
+                ">5f\n" +
+                ">69\n" +
+                ">6e7428302c203235362c20342920692c20726e\n" +
+                ">645f64\n" +
+                ">6f75626c65\n" +
+                ">28\n" +
+                ">34\n" +
+                ">2920642c\n" +
+                ">20\n" +
+                ">74\n" +
+                ">696d6573\n" +
+                ">74\n" +
+                ">616d705f\n" +
+                ">73\n" +
+                ">65717565\n" +
+                ">6e63\n" +
+                ">6528302c\n" +
+                ">31\n" +
+                ">30\n" +
+                ">303030\n" +
+                ">29\n" +
+                ">20\n" +
+                ">742c20726e\n" +
+                ">64\n" +
+                ">5f\n" +
+                ">666c6f61\n" +
+                ">74\n" +
+                ">28\n" +
+                ">34292066\n" +
+                ">2c20726e\n" +
+                ">64\n" +
+                ">5f73686f\n" +
+                ">72\n" +
+                ">74282920\n" +
+                ">5f\n" +
+                ">73\n" +
+                ">686f72742c\n" +
+                ">20\n" +
+                ">726e645f\n" +
+                ">6c\n" +
+                ">6f6e67\n" +
+                ">2830\n" +
+                ">2c20313030\n" +
+                ">30\n" +
+                ">30\n" +
+                ">303030\n" +
+                ">2c20\n" +
+                ">3529206c\n" +
+                ">2c\n" +
+                ">20\n" +
+                ">726e645f\n" +
+                ">74\n" +
+                ">696d6573\n" +
+                ">74\n" +
+                ">61\n" +
+                ">6d7028\n" +
+                ">74\n" +
+                ">6f\n" +
+                ">5f74696d\n" +
+                ">65\n" +
+                ">7374616d\n" +
+                ">70\n" +
+                ">28\n" +
+                ">27323031\n" +
+                ">35\n" +
+                ">272c27797979\n" +
+                ">79\n" +
+                ">27292c746f\n" +
+                ">5f\n" +
+                ">74696d6573\n" +
+                ">74\n" +
+                ">616d702827\n" +
+                ">32\n" +
+                ">303136\n" +
+                ">272c\n" +
+                ">277979\n" +
+                ">79\n" +
+                ">79\n" +
+                ">27292c\n" +
+                ">32\n" +
+                ">29\n" +
+                ">207473\n" +
+                ">32\n" +
+                ">2c\n" +
+                ">20726e\n" +
+                ">64\n" +
+                ">5f\n" +
+                ">627974\n" +
+                ">6528\n" +
+                ">302c31\n" +
+                ">32\n" +
+                ">37\n" +
+                ">29206262\n" +
+                ">2c\n" +
+                ">20726e\n" +
+                ">64\n" +
+                ">5f626f\n" +
+                ">6f\n" +
+                ">6c\n" +
+                ">65616e\n" +
+                ">28\n" +
+                ">29\n" +
+                ">20622c\n" +
+                ">20\n" +
+                ">72\n" +
+                ">6e645f\n" +
+                ">73\n" +
+                ">79\n" +
+                ">6d626f\n" +
+                ">6c\n" +
+                ">28342c34\n" +
+                ">2c\n" +
+                ">34\n" +
+                ">2c3229\n" +
+                ">2c\n" +
+                ">20\n" +
+                ">726e64\n" +
+                ">5f\n" +
+                ">646174\n" +
+                ">65\n" +
+                ">28\n" +
+                ">746f5f6461\n" +
+                ">74\n" +
+                ">652827\n" +
+                ">32\n" +
+                ">303135\n" +
+                ">27\n" +
+                ">2c\n" +
+                ">20\n" +
+                ">277979\n" +
+                ">7979\n" +
+                ">27292c20\n" +
+                ">74\n" +
+                ">6f\n" +
+                ">5f6461\n" +
+                ">74\n" +
+                ">652827\n" +
+                ">32\n" +
+                ">30\n" +
+                ">3136272c\n" +
+                ">2027\n" +
+                ">797979\n" +
+                ">79\n" +
+                ">27292c20\n" +
+                ">32\n" +
+                ">29\n" +
+                ">2c726e\n" +
+                ">64\n" +
+                ">5f62696e\n" +
+                ">28\n" +
+                ">31302c32\n" +
+                ">30\n" +
+                ">2c3229\n" +
+                ">20\n" +
+                ">66726f\n" +
+                ">6d\n" +
+                ">20\n" +
+                ">6c6f6e\n" +
+                ">67\n" +
+                ">5f\n" +
+                ">736571\n" +
+                ">75\n" +
+                ">656e63\n" +
+                ">65\n" +
+                ">28\n" +
+                ">35302900\n" +
+                ">00\n" +
+                ">004200\n";
         assertHexScript(script);
     }
 
@@ -5019,7 +4726,6 @@ if __name__ == "__main__":
 
     @Test
     public void testInsert2() throws Exception {
-        Assume.assumeFalse(legacyMode);
         String expectedAll = "a[INTEGER],d[TIMESTAMP],t[TIMESTAMP],d1[TIMESTAMP],t1[TIMESTAMP],t2[TIMESTAMP]\n" +
                 "0,2011-04-11 00:00:00.0,2011-04-11 14:40:54.998821,2011-04-11 14:40:54.998,2011-04-11 00:00:00.0,2011-04-11 14:40:54.998821\n" +
                 "1,2011-04-11 00:00:00.0,2011-04-11 14:40:54.999821,2011-04-11 14:40:54.999,2011-04-11 00:00:00.0,2011-04-11 14:40:54.999821\n" +
@@ -5371,7 +5077,6 @@ if __name__ == "__main__":
 
     @Test
     public void testInsertAsSelect() throws Exception {
-        Assume.assumeFalse(legacyMode);
         assertWithPgServer(CONN_AWARE_ALL, (conn, binary, mode, port) -> {
             conn.createStatement().execute("create table x(ts timestamp, col1 long) timestamp(ts) partition by day wal");
             conn.createStatement().execute("create table y(ts timestamp, col1 long) timestamp(ts) partition by day wal");
@@ -5480,9 +5185,6 @@ if __name__ == "__main__":
 
     @Test
     public void testInsertBooleans() throws Exception {
-
-        Assume.assumeFalse(legacyMode);
-        // legacy mode only works in narrow protocol cases, not all
 
         assertWithPgServer(CONN_AWARE_ALL, (connection, binary, mode, port) -> {
             connection.prepareStatement(
@@ -5961,40 +5663,6 @@ nodejs code:
     }
 
     @Test
-    public void testIntAndLongParametersWithFormatCountGreaterThanValueCount() throws Exception {
-        skipOnWalRun(); // non-partitioned table
-        Assume.assumeTrue(legacyMode); // test doesn't send a Sync, so the server is not required to respond
-        String script = ">0000006e00030000757365720078797a0064617461626173650071646200636c69656e745f656e636f64696e67005554463800446174655374796c650049534f0054696d655a6f6e65004575726f70652f4c6f6e646f6e0065787472615f666c6f61745f64696769747300320000\n" +
-                "<520000000800000003\n" +
-                ">70000000076f6800\n" +
-                "<520000000800000000530000001154696d655a6f6e6500474d5400530000001d6170706c69636174696f6e5f6e616d6500517565737444420053000000187365727665725f76657273696f6e0031312e33005300000019696e74656765725f6461746574696d6573006f6e005300000019636c69656e745f656e636f64696e670055544638004b0000000c0000003fbb8b96505a0000000549\n" +
-                ">50000000300073656c65637420782c202024312c2024322066726f6d206c6f6e675f73657175656e63652832293b000000\n" +
-                ">420000002500000003000000000000000200000001330000000a353030303030303030300000\n" +
-                "<!!";
-
-        assertHexScript(
-                NetworkFacadeImpl.INSTANCE,
-                script,
-                getStdPgWireConfigAltCreds()
-        );
-    }
-
-    @Test
-    public void testIntAndLongParametersWithFormatCountSmallerThanValueCount() throws Exception {
-        skipOnWalRun(); // non-partitioned table
-        Assume.assumeTrue(legacyMode); // test doesn't send a Sync, so the server is not required to respond
-        String script = ">0000006e00030000757365720078797a0064617461626173650071646200636c69656e745f656e636f64696e67005554463800446174655374796c650049534f0054696d655a6f6e65004575726f70652f4c6f6e646f6e0065787472615f666c6f61745f64696769747300320000\n" +
-                "<520000000800000003\n" +
-                ">70000000076f6800\n" +
-                "<520000000800000000530000001154696d655a6f6e6500474d5400530000001d6170706c69636174696f6e5f6e616d6500517565737444420053000000187365727665725f76657273696f6e0031312e33005300000019696e74656765725f6461746574696d6573006f6e005300000019636c69656e745f656e636f64696e670055544638004b0000000c0000003fbb8b96505a0000000549\n" +
-                ">50000000300073656c65637420782c202024312c2024322066726f6d206c6f6e675f73657175656e63652832293b000000\n" +
-                ">42000000230000000200000000000300000001330000000a353030303030303030300000\n" +
-                "<!!";
-
-        assertHexScript(NetworkFacadeImpl.INSTANCE, script, getStdPgWireConfigAltCreds());
-    }
-
-    @Test
     public void testIntAndLongParametersWithoutExplicitParameterTypeButOneExplicitTextFormatHex() throws Exception {
         skipOnWalRun(); // non-partitioned table
         String script = ">0000006e00030000757365720078797a0064617461626173650071646200636c69656e745f656e636f64696e67005554463800446174655374796c650049534f0054696d655a6f6e65004575726f70652f4c6f6e646f6e0065787472615f666c6f61745f64696769747300320000\n" +
@@ -6033,7 +5701,6 @@ nodejs code:
     @Test
     public void testInvalidateWriterBetweenInserts() throws Exception {
         skipOnWalRun(); // non-partitioned table
-        Assume.assumeFalse(legacyMode);
         assertWithPgServer(CONN_AWARE_ALL, (connection, binary, mode, port) -> {
             try (Statement statement = connection.createStatement()) {
                 statement.executeUpdate("create table test_batch(id long,val int)");
@@ -7088,8 +6755,6 @@ nodejs code:
 
     @Test
     public void testNamedStatementLimit() throws Exception {
-        skipInLegacyMode();
-
         assertWithPgServer(Mode.EXTENDED, true, -1, (connection, binary, mode, port) -> {
             try (Statement statement = connection.createStatement()) {
                 statement.execute("create table x as (select rnd_str() s from long_sequence(10))");
@@ -7361,27 +7026,6 @@ nodejs code:
         });
     }
 
-    @Test
-    public void testParameterTypeCountGreaterThanParameterValueCount() throws Exception {
-        Assume.assumeTrue(legacyMode); // modern server doesn't disconnect and instead ignores extra param types
-        skipOnWalRun(); // non-partitioned table
-        String script = ">0000006e00030000757365720078797a0064617461626173650071646200636c69656e745f656e636f64696e67005554463800446174655374796c650049534f0054696d655a6f6e65004575726f70652f4c6f6e646f6e0065787472615f666c6f61745f64696769747300320000\n" +
-                "<520000000800000003\n" +
-                ">70000000076f6800\n" +
-                "<520000000800000000530000001154696d655a6f6e6500474d5400530000001d6170706c69636174696f6e5f6e616d6500517565737444420053000000187365727665725f76657273696f6e0031312e33005300000019696e74656765725f6461746574696d6573006f6e005300000019636c69656e745f656e636f64696e670055544638004b0000000c0000003fbb8b96505a0000000549\n" +
-                ">5000000022005345542065787472615f666c6f61745f646967697473203d2033000000420000000c0000000000000000450000000900000000015300000004\n" +
-                "<310000000432000000044300000008534554005a0000000549\n" +
-                ">500000003700534554206170706c69636174696f6e5f6e616d65203d2027506f737467726553514c204a4442432044726976657227000000420000000c0000000000000000450000000900000000015300000004\n" +
-                "<310000000432000000044300000008534554005a0000000549\n" +
-                ">500000003b0073656c65637420782c24312c24322c24332066726f6d206c6f6e675f73657175656e63652832290000030000001700000014000002bd420000002600000003000000000000000200000001340000000331323300000004352e3433000044000000065000450000000900000000005300000004\n" +
-                "<!!";
-        assertHexScript(
-                NetworkFacadeImpl.INSTANCE,
-                script,
-                getStdPgWireConfigAltCreds()
-        );
-    }
-
     //checks that function parser error doesn't persist and affect later queries issued through the same connection
     @Test
     public void testParseErrorDoesNotCorruptConnection() throws Exception {
@@ -7431,105 +7075,6 @@ nodejs code:
                 assertResultSet("a[INTEGER],b[INTEGER]\n2,2\n", sink, result);
             }
         });
-    }
-
-    @Test
-    public void testParseMessageBadQueryTerminator() throws Exception {
-        skipOnWalRun(); // non-partitioned table
-        // test sends a malformed Parse message and doesn't send Sync, so the server isn't required to respond
-        Assume.assumeTrue(legacyMode);
-        final String script = ">0000006900030000757365720078797a006461746162617365006e6162755f61707000636c69656e745f656e636f64696e67005554463800446174655374796c650049534f0054696d655a6f6e6500474d540065787472615f666c6f61745f64696769747300320000\n" +
-                "<520000000800000003\n" +
-                ">70000000076f6800\n" +
-                "<520000000800000000530000001154696d655a6f6e6500474d5400530000001d6170706c69636174696f6e5f6e616d6500517565737444420053000000187365727665725f76657273696f6e0031312e33005300000019696e74656765725f6461746574696d6573006f6e005300000019636c69656e745f656e636f64696e670055544638004b0000000c0000003fbb8b96505a0000000549\n" +
-                ">5000000022005345542065787472615f666c6f61745f646967697473203d203308899889988998\n" +
-                "<!!";
-        assertHexScript(
-                NetworkFacadeImpl.INSTANCE,
-                script,
-                getStdPgWireConfigAltCreds()
-        );
-    }
-
-    @Test
-    public void testParseMessageBadStatementTerminator() throws Exception {
-        Assume.assumeTrue(legacyMode); // test sends a malformed packet and the server isn't required to respond
-        skipOnWalRun(); // non-partitioned table
-        final String script = ">0000006900030000757365720078797a006461746162617365006e6162755f61707000636c69656e745f656e636f64696e67005554463800446174655374796c650049534f0054696d655a6f6e6500474d540065787472615f666c6f61745f64696769747300320000\n" +
-                "<520000000800000003\n" +
-                ">70000000076f6800\n" +
-                "<520000000800000000530000001154696d655a6f6e6500474d5400530000001d6170706c69636174696f6e5f6e616d6500517565737444420053000000187365727665725f76657273696f6e0031312e33005300000019696e74656765725f6461746574696d6573006f6e005300000019636c69656e745f656e636f64696e670055544638004b0000000c0000003fbb8b96505a0000000549\n" +
-                ">5000000022555345542065787472615f666c6f61745f646967697473203d2033555555425555550c5555555555555555455555550955555555015355555504\n" +
-                "<!!";
-        assertHexScript(
-                NetworkFacadeImpl.INSTANCE,
-                script,
-                getStdPgWireConfigAltCreds()
-        );
-    }
-
-    @Test
-    public void testParseMessageNegativeParameterCount() throws Exception {
-        Assume.assumeTrue(legacyMode); // Modern server correctly responds with an error and doesn't disconnect
-        skipOnWalRun(); // non-partitioned table
-        final String script = ">0000006900030000757365720078797a006461746162617365006e6162755f61707000636c69656e745f656e636f64696e67005554463800446174655374796c650049534f0054696d655a6f6e6500474d540065787472615f666c6f61745f64696769747300320000\n" +
-                "<520000000800000003\n" +
-                ">70000000076f6800\n" +
-                "<520000000800000000530000001154696d655a6f6e6500474d5400530000001d6170706c69636174696f6e5f6e616d6500517565737444420053000000187365727665725f76657273696f6e0031312e33005300000019696e74656765725f6461746574696d6573006f6e005300000019636c69656e745f656e636f64696e670055544638004b0000000c0000003fbb8b96505a0000000549\n" +
-                ">5000000022005345542065787472615f666c6f61745f646967697473203d2033000000420000000c0000000000000000450000000900000000015300000004\n" +
-                "<310000000432000000044300000008534554005a0000000549\n" +
-                ">500000003700534554206170706c69636174696f6e5f6e616d65203d2027506f737467726553514c204a4442432044726976657227000000420000000c0000000000000000450000000900000000015300000004\n" +
-                "<310000000432000000044300000008534554005a0000000549\n" +
-                ">50000000cd0073656c65637420782c24312c24322c24332c24342c24352c24362c24372c24382c24392c2431302c2431312c2431322c2431332c2431342c2431352c2431362c2431372c2431382c2431392c2432302c2432312c2432322066726f6d206c6f6e675f73657175656e636528352900fefe0000001700000014000002bc000002bd0000001500000010000004130000041300000000000000000000001700000014000002bc000002bd000000150000001000000413000004130000043a000000000000045a000004a0420000012c0000001600010001000100010001000000000000000000000001000100010001000100000000000000010000000000000016000000040000000400000008000000000000007b0000000440adc28f000000083fe22c27a63736ce00000002005b00000004545255450000000568656c6c6f0000001dd0b3d180d183d0bfd0bfd0b020d182d183d180d0b8d181d182d0bed0b20000000e313937302d30312d3031202b30300000001a313937302d30382d32302031313a33333a32302e3033332b3030ffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffff0000001a313937302d30312d30312030303a30353a30302e3031312b30300000001a313937302d30312d30312030303a30383a32302e3032332b3030000044000000065000450000000900000000005300000004\n" +
-                "<!!";
-        assertHexScript(
-                NetworkFacadeImpl.INSTANCE,
-                script,
-                getStdPgWireConfigAltCreds()
-        );
-    }
-
-    @Test
-    public void testParseMessageTruncatedAtParameter() throws Exception {
-        Assume.assumeTrue(legacyMode); // Modern server correctly responds with an error and doesn't disconnect
-        skipOnWalRun(); // non-partitioned table
-        final String script = ">0000006900030000757365720078797a006461746162617365006e6162755f61707000636c69656e745f656e636f64696e67005554463800446174655374796c650049534f0054696d655a6f6e6500474d540065787472615f666c6f61745f64696769747300320000\n" +
-                "<520000000800000003\n" +
-                ">70000000076f6800\n" +
-                "<520000000800000000530000001154696d655a6f6e6500474d5400530000001d6170706c69636174696f6e5f6e616d6500517565737444420053000000187365727665725f76657273696f6e0031312e33005300000019696e74656765725f6461746574696d6573006f6e005300000019636c69656e745f656e636f64696e670055544638004b0000000c0000003fbb8b96505a0000000549\n" +
-                ">5000000022005345542065787472615f666c6f61745f646967697473203d2033000000420000000c0000000000000000450000000900000000015300000004\n" +
-                "<310000000432000000044300000008534554005a0000000549\n" +
-                ">500000003700534554206170706c69636174696f6e5f6e616d65203d2027506f737467726553514c204a4442432044726976657227000000420000000c0000000000000000450000000900000000015300000004\n" +
-                "<310000000432000000044300000008534554005a0000000549\n" +
-                ">50000000cd0073656c65637420782c24312c24322c24332c24342c24352c24362c24372c24382c24392c2431302c2431312c2431322c2431332c2431342c2431352c2431362c2431372c2431382c2431392c2432302c2432312c2432322066726f6d206c6f6e675f73657175656e63652835290000260000001700000014000002bc000002bd0000001500000010000004130000041300000000000000000000001700000014000002bc000002bd000000150000001000000413000004130000043a000000000000045a000004a0420000012c0000001600010001000100010001000000000000000000000001000100010001000100000000000000010000000000000016000000040000000400000008000000000000007b0000000440adc28f000000083fe22c27a63736ce00000002005b00000004545255450000000568656c6c6f0000001dd0b3d180d183d0bfd0bfd0b020d182d183d180d0b8d181d182d0bed0b20000000e313937302d30312d3031202b30300000001a313937302d30382d32302031313a33333a32302e3033332b3030ffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffff0000001a313937302d30312d30312030303a30353a30302e3031312b30300000001a313937302d30312d30312030303a30383a32302e3032332b3030000044000000065000450000000900000000005300000004\n" +
-                "<!!";
-        assertHexScript(
-                NetworkFacadeImpl.INSTANCE,
-                script,
-                getStdPgWireConfigAltCreds()
-        );
-    }
-
-    @Test
-    public void testParseMessageTruncatedAtParameterCount() throws Exception {
-        Assume.assumeTrue(legacyMode); // Since the test doesn't send a Sync message, server isn't required to respond
-        skipOnWalRun(); // non-partitioned table
-        final String script = ">0000006900030000757365720078797a006461746162617365006e6162755f61707000636c69656e745f656e636f64696e67005554463800446174655374796c650049534f0054696d655a6f6e6500474d540065787472615f666c6f61745f64696769747300320000\n" +
-                "<520000000800000003\n" +
-                ">70000000076f6800\n" +
-                "<520000000800000000530000001154696d655a6f6e6500474d5400530000001d6170706c69636174696f6e5f6e616d6500517565737444420053000000187365727665725f76657273696f6e0031312e33005300000019696e74656765725f6461746574696d6573006f6e005300000019636c69656e745f656e636f64696e670055544638004b0000000c0000003fbb8b96505a0000000549\n" +
-                ">5000000022005345542065787472615f666c6f61745f646967697473203d2033000000420000000c0000000000000000450000000900000000015300000004\n" +
-                "<310000000432000000044300000008534554005a0000000549\n" +
-                ">500000003700534554206170706c69636174696f6e5f6e616d65203d2027506f737467726553514c204a4442432044726976657227000000420000000c0000000000000000450000000900000000015300000004\n" +
-                "<310000000432000000044300000008534554005a0000000549\n" +
-                // II
-                ">50000000740073656c65637420782c24312c24322c24332c24342c24352c24362c24372c24382c24392c2431302c2431312c2431322c2431332c2431342c2431352c2431362c2431372c2431382c2431392c2432302c2432312c2432322066726f6d206c6f6e675f73657175656e63652835290000160000001700000014000002bc000002bd0000001500000010000004130000041300000000000000000000001700000014000002bc000002bd000000150000001000000413000004130000043a000000000000045a000004a0420000012c0000001600010001000100010001000000000000000000000001000100010001000100000000000000010000000000000016000000040000000400000008000000000000007b0000000440adc28f000000083fe22c27a63736ce00000002005b00000004545255450000000568656c6c6f0000001dd0b3d180d183d0bfd0bfd0b020d182d183d180d0b8d181d182d0bed0b20000000e313937302d30312d3031202b30300000001a313937302d30382d32302031313a33333a32302e3033332b3030ffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffff0000001a313937302d30312d30312030303a30353a30302e3031312b30300000001a313937302d30312d30312030303a30383a32302e3032332b3030000044000000065000450000000900000000005300000004\n" +
-                "<!!";
-        assertHexScript(
-                NetworkFacadeImpl.INSTANCE,
-                script,
-                getStdPgWireConfigAltCreds()
-        );
     }
 
     @Test
@@ -7738,92 +7283,18 @@ nodejs code:
     public void testPreparedStatementParamBadByte() throws Exception {
         skipOnWalRun(); // non-partitioned table
         final String script;
-        if (legacyMode) {
-            // old code will disconnect if string bind variable value has invalid UTF8 encoding
-            script = ">0000006b00030000757365720061646d696e006461746162617365006e6162755f61707000636c69656e745f656e636f64696e67005554463800446174655374796c650049534f0054696d655a6f6e6500474d540065787472615f666c6f61745f64696769747300320000\n" +
-                    "<520000000800000003\n" +
-                    ">700000000a717565737400\n" +
-                    "<520000000800000000530000001154696d655a6f6e6500474d5400530000001d6170706c69636174696f6e5f6e616d6500517565737444420053000000187365727665725f76657273696f6e0031312e33005300000019696e74656765725f6461746574696d6573006f6e005300000019636c69656e745f656e636f64696e670055544638004b0000000c0000003fbb8b96505a0000000549\n" +
-                    ">5000000022005345542065787472615f666c6f61745f646967697473203d2033000000420000000c0000000000000000450000000900000000015300000004\n" +
-                    "<310000000432000000044300000008534554005a0000000549\n" +
-                    ">500000003700534554206170706c69636174696f6e5f6e616d65203d2027506f737467726553514c204a4442432044726976657227000000420000000c0000000000000000450000000900000000015300000004\n" +
-                    "<310000000432000000044300000008534554005a0000000549\n" +
-                    ">50000000cd0073656c65637420782c24312c24322c24332c24342c24352c24362c24372c24382c24392c2431302c2431312c2431322c2431332c2431342c2431352c2431362c2431372c2431382c2431392c2432302c2432312c2432322066726f6d206c6f6e675f73657175656e63652835290000160000001700000014000002bd000002bd0000001500000010000004130000041300000000000000000000001700000014000002bc000002bd000000150000001000000413000004130000043a000000000000045a000004a04200000123000000160000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001600000001340000000331323300000004352e343300000007302e353637383900000002993100000004545255450000000568656c6c6f0000001dd0b3d180d183d0bfd0bfd0b020d182d183d180d0b8d181d182d0bed0b20000000e313937302d30312d3031202b30300000001a313937302d30382d32302031313a33333a32302e3033332b3030ffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffff0000001a313937302d30312d30312030303a30353a30302e3031312b30300000001a313937302d30312d30312030303a30383a32302e3032332b3030000044000000065000450000000900000000005300000004\n" +
-                    "<!!";
-        } else {
-            // the modern client will produce actionable error message to the client
-            script = ">0000006b00030000757365720061646d696e006461746162617365006e6162755f61707000636c69656e745f656e636f64696e67005554463800446174655374796c650049534f0054696d655a6f6e6500474d540065787472615f666c6f61745f64696769747300320000\n" +
-                    "<520000000800000003\n" +
-                    ">700000000a717565737400\n" +
-                    "<520000000800000000530000001154696d655a6f6e6500474d5400530000001d6170706c69636174696f6e5f6e616d6500517565737444420053000000187365727665725f76657273696f6e0031312e33005300000019696e74656765725f6461746574696d6573006f6e005300000019636c69656e745f656e636f64696e670055544638004b0000000c0000003fbb8b96505a0000000549\n" +
-                    ">5000000022005345542065787472615f666c6f61745f646967697473203d2033000000420000000c0000000000000000450000000900000000015300000004\n" +
-                    "<310000000432000000044300000008534554005a0000000549\n" +
-                    ">500000003700534554206170706c69636174696f6e5f6e616d65203d2027506f737467726553514c204a4442432044726976657227000000420000000c0000000000000000450000000900000000015300000004\n" +
-                    "<310000000432000000044300000008534554005a0000000549\n" +
-                    ">50000000cd0073656c65637420782c24312c24322c24332c24342c24352c24362c24372c24382c24392c2431302c2431312c2431322c2431332c2431342c2431352c2431362c2431372c2431382c2431392c2432302c2432312c2432322066726f6d206c6f6e675f73657175656e63652835290000160000001700000014000002bd000002bd0000001500000010000004130000041300000000000000000000001700000014000002bc000002bd000000150000001000000413000004130000043a000000000000045a000004a04200000123000000160000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001600000001340000000331323300000004352e343300000007302e353637383900000002993100000004545255450000000568656c6c6f0000001dd0b3d180d183d0bfd0bfd0b020d182d183d180d0b8d181d182d0bed0b20000000e313937302d30312d3031202b30300000001a313937302d30382d32302031313a33333a32302e3033332b3030ffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffff0000001a313937302d30312d30312030303a30353a30302e3031312b30300000001a313937302d30312d30312030303a30383a32302e3032332b3030000044000000065000450000000900000000005300000004\n" +
-                    "<4500000050433030303030004d696e76616c6964205554463820656e636f64696e6720666f7220737472696e672076616c7565205b7661726961626c65496e6465783d345d00534552524f5200503100005a0000000549\n";
-        }
+        // the modern client will produce actionable error message to the client
+        script = ">0000006b00030000757365720061646d696e006461746162617365006e6162755f61707000636c69656e745f656e636f64696e67005554463800446174655374796c650049534f0054696d655a6f6e6500474d540065787472615f666c6f61745f64696769747300320000\n" +
+                "<520000000800000003\n" +
+                ">700000000a717565737400\n" +
+                "<520000000800000000530000001154696d655a6f6e6500474d5400530000001d6170706c69636174696f6e5f6e616d6500517565737444420053000000187365727665725f76657273696f6e0031312e33005300000019696e74656765725f6461746574696d6573006f6e005300000019636c69656e745f656e636f64696e670055544638004b0000000c0000003fbb8b96505a0000000549\n" +
+                ">5000000022005345542065787472615f666c6f61745f646967697473203d2033000000420000000c0000000000000000450000000900000000015300000004\n" +
+                "<310000000432000000044300000008534554005a0000000549\n" +
+                ">500000003700534554206170706c69636174696f6e5f6e616d65203d2027506f737467726553514c204a4442432044726976657227000000420000000c0000000000000000450000000900000000015300000004\n" +
+                "<310000000432000000044300000008534554005a0000000549\n" +
+                ">50000000cd0073656c65637420782c24312c24322c24332c24342c24352c24362c24372c24382c24392c2431302c2431312c2431322c2431332c2431342c2431352c2431362c2431372c2431382c2431392c2432302c2432312c2432322066726f6d206c6f6e675f73657175656e63652835290000160000001700000014000002bd000002bd0000001500000010000004130000041300000000000000000000001700000014000002bc000002bd000000150000001000000413000004130000043a000000000000045a000004a04200000123000000160000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001600000001340000000331323300000004352e343300000007302e353637383900000002993100000004545255450000000568656c6c6f0000001dd0b3d180d183d0bfd0bfd0b020d182d183d180d0b8d181d182d0bed0b20000000e313937302d30312d3031202b30300000001a313937302d30382d32302031313a33333a32302e3033332b3030ffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffff0000001a313937302d30312d30312030303a30353a30302e3031312b30300000001a313937302d30312d30312030303a30383a32302e3032332b3030000044000000065000450000000900000000005300000004\n" +
+                "<4500000050433030303030004d696e76616c6964205554463820656e636f64696e6720666f7220737472696e672076616c7565205b7661726961626c65496e6465783d345d00534552524f5200503100005a0000000549\n";
         assertHexScript(NetworkFacadeImpl.INSTANCE, script, new Port0PGWireConfiguration());
-    }
-
-    @Test
-    public void testPreparedStatementParamBadInt() throws Exception {
-        Assume.assumeTrue(legacyMode); // modern server correctly responds with an error and doesn't disconnect
-        skipOnWalRun(); // non-partitioned table
-        assertHexScript(
-                NetworkFacadeImpl.INSTANCE,
-                ">0000006b00030000757365720061646d696e006461746162617365006e6162755f61707000636c69656e745f656e636f64696e67005554463800446174655374796c650049534f0054696d655a6f6e6500474d540065787472615f666c6f61745f64696769747300320000\n" +
-                        "<520000000800000003\n" +
-                        ">700000000a717565737400\n" +
-                        "<520000000800000000530000001154696d655a6f6e6500474d5400530000001d6170706c69636174696f6e5f6e616d6500517565737444420053000000187365727665725f76657273696f6e0031312e33005300000019696e74656765725f6461746574696d6573006f6e005300000019636c69656e745f656e636f64696e670055544638004b0000000c0000003fbb8b96505a0000000549\n" +
-                        ">5000000022005345542065787472615f666c6f61745f646967697473203d2033000000420000000c0000000000000000450000000900000000015300000004\n" +
-                        "<310000000432000000044300000008534554005a0000000549\n" +
-                        ">500000003700534554206170706c69636174696f6e5f6e616d65203d2027506f737467726553514c204a4442432044726976657227000000420000000c0000000000000000450000000900000000015300000004\n" +
-                        "<310000000432000000044300000008534554005a0000000549\n" +
-                        ">50000000cd0073656c65637420782c24312c24322c24332c24342c24352c24362c24372c24382c24392c2431302c2431312c2431322c2431332c2431342c2431352c2431362c2431372c2431382c2431392c2432302c2432312c2432322066726f6d206c6f6e675f73657175656e63652835290000160000001700000014000002bd000002bd0000001500000010000004130000041300000000000000000000001700000014000002bc000002bd000000150000001000000413000004130000043a000000000000045a000004a04200000123000000160000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001600000001fc0000000331323300000004352e343300000007302e353637383900000002393100000004545255450000000568656c6c6f0000001dd0b3d180d183d0bfd0bfd0b020d182d183d180d0b8d181d182d0bed0b20000000e313937302d30312d3031202b30300000001a313937302d30382d32302031313a33333a32302e3033332b3030ffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffff0000001a313937302d30312d30312030303a30353a30302e3031312b30300000001a313937302d30312d30312030303a30383a32302e3032332b3030000044000000065000450000000900000000005300000004\n" +
-                        "<!!",
-                new Port0PGWireConfiguration()
-        );
-    }
-
-    @Test
-    public void testPreparedStatementParamBadLong() throws Exception {
-        Assume.assumeTrue(legacyMode); // modern server correctly responds with an error and doesn't disconnect
-        skipOnWalRun(); // non-partitioned table
-        assertHexScript(
-                NetworkFacadeImpl.INSTANCE,
-                ">0000006b00030000757365720061646d696e006461746162617365006e6162755f61707000636c69656e745f656e636f64696e67005554463800446174655374796c650049534f0054696d655a6f6e6500474d540065787472615f666c6f61745f64696769747300320000\n" +
-                        "<520000000800000003\n" +
-                        ">700000000a717565737400\n" +
-                        "<520000000800000000530000001154696d655a6f6e6500474d5400530000001d6170706c69636174696f6e5f6e616d6500517565737444420053000000187365727665725f76657273696f6e0031312e33005300000019696e74656765725f6461746574696d6573006f6e005300000019636c69656e745f656e636f64696e670055544638004b0000000c0000003fbb8b96505a0000000549\n" +
-                        ">5000000022005345542065787472615f666c6f61745f646967697473203d2033000000420000000c0000000000000000450000000900000000015300000004\n" +
-                        "<310000000432000000044300000008534554005a0000000549\n" +
-                        ">500000003700534554206170706c69636174696f6e5f6e616d65203d2027506f737467726553514c204a4442432044726976657227000000420000000c0000000000000000450000000900000000015300000004\n" +
-                        "<310000000432000000044300000008534554005a0000000549\n" +
-                        ">50000000cd0073656c65637420782c24312c24322c24332c24342c24352c24362c24372c24382c24392c2431302c2431312c2431322c2431332c2431342c2431352c2431362c2431372c2431382c2431392c2432302c2432312c2432322066726f6d206c6f6e675f73657175656e63652835290000160000001700000014000002bd000002bd0000001500000010000004130000041300000000000000000000001700000014000002bc000002bd000000150000001000000413000004130000043a000000000000045a000004a04200000123000000160000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001600000001340000000331b23300000004352e343300000007302e353637383900000002393100000004545255450000000568656c6c6f0000001dd0b3d180d183d0bfd0bfd0b020d182d183d180d0b8d181d182d0bed0b20000000e313937302d30312d3031202b30300000001a313937302d30382d32302031313a33333a32302e3033332b3030ffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffff0000001a313937302d30312d30312030303a30353a30302e3031312b30300000001a313937302d30312d30312030303a30383a32302e3032332b3030000044000000065000450000000900000000005300000004\n" +
-                        "<!!",
-                new Port0PGWireConfiguration()
-        );
-    }
-
-    @Test
-    public void testPreparedStatementParamValueLengthOverflow() throws Exception {
-        Assume.assumeTrue(legacyMode); // modern server correctly responds with an error and doesn't disconnect
-        skipOnWalRun(); // non-partitioned table
-        assertHexScript(
-                NetworkFacadeImpl.INSTANCE,
-                ">0000006b00030000757365720061646d696e006461746162617365006e6162755f61707000636c69656e745f656e636f64696e67005554463800446174655374796c650049534f0054696d655a6f6e6500474d540065787472615f666c6f61745f64696769747300320000\n" +
-                        "<520000000800000003\n" +
-                        ">700000000a717565737400\n" +
-                        "<520000000800000000530000001154696d655a6f6e6500474d5400530000001d6170706c69636174696f6e5f6e616d6500517565737444420053000000187365727665725f76657273696f6e0031312e33005300000019696e74656765725f6461746574696d6573006f6e005300000019636c69656e745f656e636f64696e670055544638004b0000000c0000003fbb8b96505a0000000549\n" +
-                        ">5000000022005345542065787472615f666c6f61745f646967697473203d2033000000420000000c0000000000000000450000000900000000015300000004\n" +
-                        "<310000000432000000044300000008534554005a0000000549\n" +
-                        ">500000003700534554206170706c69636174696f6e5f6e616d65203d2027506f737467726553514c204a4442432044726976657227000000420000000c0000000000000000450000000900000000015300000004\n" +
-                        "<310000000432000000044300000008534554005a0000000549\n" +
-                        ">50000000cd0073656c65637420782c24312c24322c24332c24342c24352c24362c24372c24382c24392c2431302c2431312c2431322c2431332c2431342c2431352c2431362c2431372c2431382c2431392c2432302c2432312c2432322066726f6d206c6f6e675f73657175656e63652835290000160000001700000014000002bd000002bd0000001500000010000004130000041300000000000000000000001700000014000002bc000002bd000000150000001000000413000004130000043a000000000000045a000004a04200000123000000160000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001600000001340000333331b23300000004352e343300000007302e353637383900000002393100000004545255450000000568656c6c6f0000001dd0b3d180d183d0bfd0bfd0b020d182d183d180d0b8d181d182d0bed0b20000000e313937302d30312d3031202b30300000001a313937302d30382d32302031313a33333a32302e3033332b3030ffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffff0000001a313937302d30312d30312030303a30353a30302e3031312b30300000001a313937302d30312d30312030303a30383a32302e3032332b3030000044000000065000450000000900000000005300000004\n" +
-                        "<!!",
-                new Port0PGWireConfiguration()
-        );
     }
 
     @Test
@@ -7834,11 +7305,6 @@ nodejs code:
                 @Override
                 public int getWorkerCount() {
                     return 4;
-                }
-
-                @Override
-                public boolean isLegacyModeEnabled() {
-                    return legacyMode;
                 }
             };
 
@@ -7891,14 +7357,6 @@ nodejs code:
                     statement.setTimestamp(21, new PGTimestamp(500023, new GregorianCalendar()));
                     statement.setTimestamp(22, null);
 
-                    // in legacy record cursor metadata is decided before the variable types are known
-                    final String legacyExpected = "x[BIGINT],$1[VARCHAR],$2[VARCHAR],$3[VARCHAR],$4[VARCHAR],$5[VARCHAR],$6[VARCHAR],$7[VARCHAR],$8[VARCHAR],$9[VARCHAR],$10[VARCHAR],$11[VARCHAR],$12[VARCHAR],$13[VARCHAR],$14[VARCHAR],$15[VARCHAR],$16[VARCHAR],$17[VARCHAR],$18[VARCHAR],$19[VARCHAR],$20[VARCHAR],$21[VARCHAR],$22[VARCHAR]\n" +
-                            "1,4,123,5.43,0.56789,91,TRUE,hello,группа туристов,1970-01-01 +00,1970-08-20 11:33:20.033+00,null,null,null,null,0,null,null,null,null,1970-01-01 00:05:00.011+00,1970-01-01 00:08:20.023+00,null\n" +
-                            "2,4,123,5.43,0.56789,91,TRUE,hello,группа туристов,1970-01-01 +00,1970-08-20 11:33:20.033+00,null,null,null,null,0,null,null,null,null,1970-01-01 00:05:00.011+00,1970-01-01 00:08:20.023+00,null\n" +
-                            "3,4,123,5.43,0.56789,91,TRUE,hello,группа туристов,1970-01-01 +00,1970-08-20 11:33:20.033+00,null,null,null,null,0,null,null,null,null,1970-01-01 00:05:00.011+00,1970-01-01 00:08:20.023+00,null\n" +
-                            "4,4,123,5.43,0.56789,91,TRUE,hello,группа туристов,1970-01-01 +00,1970-08-20 11:33:20.033+00,null,null,null,null,0,null,null,null,null,1970-01-01 00:05:00.011+00,1970-01-01 00:08:20.023+00,null\n" +
-                            "5,4,123,5.43,0.56789,91,TRUE,hello,группа туристов,1970-01-01 +00,1970-08-20 11:33:20.033+00,null,null,null,null,0,null,null,null,null,1970-01-01 00:05:00.011+00,1970-01-01 00:08:20.023+00,null\n";
-
                     // modern uses metadata from the PARSE message
                     final String modernExpected = "x[BIGINT],$1[INTEGER],$2[BIGINT],$3[REAL],$4[DOUBLE],$5[SMALLINT],$6[BIT],$7[VARCHAR],$8[VARCHAR],$9[VARCHAR],$10[VARCHAR],$11[INTEGER],$12[BIGINT],$13[REAL],$14[DOUBLE],$15[SMALLINT],$16[BIT],$17[VARCHAR],$18[VARCHAR],$19[TIMESTAMP],$20[TIMESTAMP],$21[TIMESTAMP],$22[VARCHAR]\n" +
                             "1,4,123,5.430,0.56789,91,true,hello,группа туристов,1970-01-01 +00,1970-08-20 11:33:20.033+00,null,null,null,null,0,false,null,null,null,1970-01-01 00:05:00.011,1970-01-01 00:08:20.023,null\n" +
@@ -7906,13 +7364,12 @@ nodejs code:
                             "3,4,123,5.430,0.56789,91,true,hello,группа туристов,1970-01-01 +00,1970-08-20 11:33:20.033+00,null,null,null,null,0,false,null,null,null,1970-01-01 00:05:00.011,1970-01-01 00:08:20.023,null\n" +
                             "4,4,123,5.430,0.56789,91,true,hello,группа туристов,1970-01-01 +00,1970-08-20 11:33:20.033+00,null,null,null,null,0,false,null,null,null,1970-01-01 00:05:00.011,1970-01-01 00:08:20.023,null\n" +
                             "5,4,123,5.430,0.56789,91,true,hello,группа туристов,1970-01-01 +00,1970-08-20 11:33:20.033+00,null,null,null,null,0,false,null,null,null,1970-01-01 00:05:00.011,1970-01-01 00:08:20.023,null\n";
-                    final String expected = legacyMode ? legacyExpected : modernExpected;
 
                     StringSink sink = new StringSink();
                     for (int i = 0; i < 10000; i++) {
                         sink.clear();
                         ResultSet rs = statement.executeQuery();
-                        assertResultSet(expected, sink, rs);
+                        assertResultSet(modernExpected, sink, rs);
                         rs.close();
                     }
                     connection.close();
@@ -7997,7 +7454,6 @@ nodejs code:
         sendBufferSize = 1024;
         recvBufferSize = 1024;
         // there is a bug in legacy mode somewhere
-        Assume.assumeFalse(legacyMode);
         assertWithPgServer(CONN_AWARE_ALL, (connection, binary, mode, port) -> {
             PreparedStatement statement = connection.prepareStatement("select x, ? as \"$1\",? as \"$2\",? as \"$3\",? as \"$4\"," +
                     "? as \"$5\",? as \"$6\",? as \"$7\",? as \"$8\",? as \"$9\",? as \"$10\",? as \"$11\",? as \"$12\",? as \"$13\"," +
@@ -8136,10 +7592,6 @@ nodejs code:
 
     @Test
     public void testPreparedStatementWithBindVariablesSetWrongOnSameConnection() throws Exception {
-
-        // this test will fail in Legacy mode due to a bug in the legacy server code
-        Assume.assumeFalse(legacyMode);
-
         assertWithPgServer(CONN_AWARE_EXTENDED, (connection, binary, mode, port) -> {
             try (PreparedStatement statement = connection.prepareStatement(createDatesTblStmt)) {
                 statement.execute();
@@ -8215,7 +7667,6 @@ nodejs code:
 
     @Test
     public void testPreparedStatementWithSystimestampFunction() throws Exception {
-        Assume.assumeFalse(legacyMode);
         assertWithPgServer(CONN_AWARE_ALL, (connection, binary, mode, port) -> {
             try (PreparedStatement statement = connection.prepareStatement(
                     "create table xts (ts timestamp) timestamp(ts)")) {
@@ -8336,7 +7787,6 @@ nodejs code:
 
     @Test
     public void testQueryAgainstIndexedSymbol() throws Exception {
-        Assume.assumeFalse(legacyMode);
         Assume.assumeTrue(walEnabled);
         final String[] values = {"'5'", "null", "'5' || ''", "replace(null, 'A', 'A')", "?5", "?null"};
         final CharSequenceObjHashMap<String> valMap = new CharSequenceObjHashMap<>();
@@ -8463,7 +7913,6 @@ nodejs code:
 
     @Test
     public void testQueryCountMetrics() throws Exception {
-        Assume.assumeFalse(legacyMode); // only modern PGWire server supports query counters
         skipOnWalRun(); // non-partitioned table
         assertWithPgServer(CONN_AWARE_ALL, (connection, binary, mode, port) -> {
             execute("create table x as (select x id from long_sequence(10))");
@@ -8537,7 +7986,6 @@ nodejs code:
     @Test
     public void testQueryEventuallySucceedsOnDataUnavailableEventNeverFired() throws Exception {
         Assume.assumeFalse(walEnabled);
-        Assume.assumeFalse(legacyMode);
         maxQueryTime = 100;
         assertWithPgServer(CONN_AWARE_ALL, (connection, binary, mode, port) -> {
             AtomicReference<SuspendEvent> eventRef = new AtomicReference<>();
@@ -8715,7 +8163,6 @@ nodejs code:
     @Test
     public void testQueryTimeoutModern() throws Exception {
         skipOnWalRun(); // non-partitioned table
-        Assume.assumeFalse(legacyMode);
         maxQueryTime = 100;
         assertWithPgServer(CONN_AWARE_ALL, (connection, binary, mode, port) -> {
             try (final PreparedStatement statement = connection.prepareStatement("select sleep(120000)")) {
@@ -9127,7 +8574,6 @@ nodejs code:
     @Test
     public void testRunQueryAfterCancellingPreviousInTheSameConnection() throws Exception {
         // yet enother bug in legacy code, this test would be unstable
-        Assume.assumeFalse(legacyMode);
         assertWithPgServer(CONN_AWARE_ALL, (connection, binary, mode, port) -> {
             execute("create table if not exists tab as " +
                     "(select x::timestamp ts, " +
@@ -9340,7 +8786,6 @@ create table tab as (
      */
     public void testSelectAllTypesFromAsyncPG() throws Exception {
         skipOnWalRun(); // non-partitioned table
-        skipInLegacyMode(); // legacy impl does not send the right type modifier for CHAR columns
         execute("create table tab as (\n" +
                 "    select\n" +
                 "        rnd_byte() b,\n" +
@@ -10172,7 +9617,6 @@ create table tab as (
     @Test
     public void testSmallSendBufferBigColumnValueNotEnoughSpace2() throws Exception {
         // Here we test user-friendly error messages which are only present in the modern PGWire.
-        Assume.assumeFalse(legacyMode);
         Assume.assumeFalse(walEnabled);
 
         recvBufferSize = 1024;
@@ -10282,8 +9726,6 @@ create table tab as (
 
     @Test
     public void testSmallSendBufferForRowDescription() throws Exception {
-        // legacy code fails this test in SIMPLE protocol, modern code does not
-        Assume.assumeFalse(legacyMode);
         // WAL is irrelevant here, we are checking result sending code path
         skipOnWalRun();
         String sqlCreate = "create table x as (" +
@@ -10338,8 +9780,6 @@ create table tab as (
 
     @Test
     public void testSmallSendBufferLargeErrorMessage() throws Exception {
-        // We want to test behavior of the modern PGWire only.
-        Assume.assumeFalse(legacyMode);
         Assume.assumeFalse(walEnabled);
 
         sendBufferSize = 256;
@@ -10379,8 +9819,6 @@ create table tab as (
 
     @Test
     public void testSmallSendBufferWideRecordPermute() throws Exception {
-        // legacy code fails this test, modern code does not
-        Assume.assumeFalse(legacyMode);
         // WAL is irrelevant here, we are checking result sending code path
         skipOnWalRun();
 
@@ -10564,33 +10002,17 @@ create table tab as (
     public void testSyncAfterLoginSendsRNQ() throws Exception {
         skipOnWalRun();
         final String script;
-        if (legacyMode) {
-            script = ">0000000804d2162f\n" +
-                    "<4e\n" +
-                    ">0000006b00030000757365720061646d696e0064617461626173650071646200446174655374796c650049534f2c204d445900636c69656e745f656e636f64696e6700555446380054696d655a6f6e65005554430065787472615f666c6f61745f64696769747300330000\n" +
-                    "<520000000800000003\n" +
-                    ">700000000a717565737400\n" +
-                    "<520000000800000000530000001154696d655a6f6e6500474d5400530000001d6170706c69636174696f6e5f6e616d6500517565737444420053000000187365727665725f76657273696f6e0031312e33005300000019696e74656765725f6461746574696d6573006f6e005300000019636c69656e745f656e636f64696e670055544638004b0000000c0000003fbb8b96505a0000000549\n" +
-                    ">5300000004\n" +
-                    "<5a0000000549\n" +
-                    ">500000003373716c785f735f310053454c4543542024312066726f6d206c6f6e675f73657175656e636528322900000100000017440000000e5373716c785f735f31005300000004\n" +
-                    "<3100000004740000000a000100000017540000001b000124310000000000000100000413ffffffffffff00005a0000000549\n" +
-                    ">42000000200073716c785f735f310000010001000100000004000000010001000145000000090000000000430000000650005300000004\n" +
-                    "<3200000004440000000b00010000000131440000000b00010000000131430000000d53454c45435420320033000000045a0000000549\n" +
-                    ">5800000004";
-        } else {
-            script =
-                    ">0000006b00030000757365720061646d696e0064617461626173650071646200446174655374796c650049534f2c204d445900636c69656e745f656e636f64696e6700555446380054696d655a6f6e65005554430065787472615f666c6f61745f64696769747300320000\n" +
-                            "<520000000800000003\n" +
-                            ">700000000a717565737400\n" +
-                            "<520000000800000000530000001154696d655a6f6e6500474d5400530000001d6170706c69636174696f6e5f6e616d6500517565737444420053000000187365727665725f76657273696f6e0031312e33005300000019696e74656765725f6461746574696d6573006f6e005300000019636c69656e745f656e636f64696e670055544638004b0000000c0000003fbb8b96505a0000000549\n" +
-                            ">5300000004\n" +
-                            "<5a0000000549\n" +
-                            ">500000003373716c785f735f310053454c4543542024312066726f6d206c6f6e675f73657175656e636528322900000100000017440000000e5373716c785f735f31005300000004\n" +
-                            "<3100000004740000000a000100000017540000001b0001243100000000000001000000170004ffffffff00005a0000000549\n" +
-                            ">42000000200073716c785f735f310000010001000100000004000000010001000145000000090000000000430000000650005300000004\n" +
-                            "<3200000004440000000e00010000000400000001440000000e00010000000400000001430000000d53454c45435420320033000000045a0000000549\n";
-        }
+        script =
+                ">0000006b00030000757365720061646d696e0064617461626173650071646200446174655374796c650049534f2c204d445900636c69656e745f656e636f64696e6700555446380054696d655a6f6e65005554430065787472615f666c6f61745f64696769747300320000\n" +
+                        "<520000000800000003\n" +
+                        ">700000000a717565737400\n" +
+                        "<520000000800000000530000001154696d655a6f6e6500474d5400530000001d6170706c69636174696f6e5f6e616d6500517565737444420053000000187365727665725f76657273696f6e0031312e33005300000019696e74656765725f6461746574696d6573006f6e005300000019636c69656e745f656e636f64696e670055544638004b0000000c0000003fbb8b96505a0000000549\n" +
+                        ">5300000004\n" +
+                        "<5a0000000549\n" +
+                        ">500000003373716c785f735f310053454c4543542024312066726f6d206c6f6e675f73657175656e636528322900000100000017440000000e5373716c785f735f31005300000004\n" +
+                        "<3100000004740000000a000100000017540000001b0001243100000000000001000000170004ffffffff00005a0000000549\n" +
+                        ">42000000200073716c785f735f310000010001000100000004000000010001000145000000090000000000430000000650005300000004\n" +
+                        "<3200000004440000000e00010000000400000001440000000e00010000000400000001430000000d53454c45435420320033000000045a0000000549\n";
         assertHexScript(NetworkFacadeImpl.INSTANCE, script, new Port0PGWireConfiguration());
     }
 
@@ -10626,7 +10048,6 @@ create table tab as (
     @Test
     public void testTableReferenceOutOfDate() throws Exception {
         Assume.assumeFalse(walEnabled);
-        Assume.assumeFalse(legacyMode);
         assertWithPgServer(CONN_AWARE_ALL, (connection, binary, mode, port) -> {
             final String query = "select * from test_table_reference_out_of_date();";
             try (
@@ -10896,7 +10317,6 @@ create table tab as (
     @Test
     public void testUndefinedBindVariableInSymbol() throws Exception {
         // @Ignore("ERROR: bind variable at 0 is defined as unknown and cannot accept STRING")
-        Assume.assumeFalse(legacyMode);
         final String[] values = {"'5'", "null", "'5' || ''", "replace(null, 'A', 'A')", "?5", "?null"};
         final CharSequenceObjHashMap<String> valMap = new CharSequenceObjHashMap<>();
         valMap.put("5", "5");
