+++ core/src/test/java/io/questdb/test/cutlass/pgwire/PGMemoryLimitTest.java
@@ -27,28 +27,15 @@ package io.questdb.test.cutlass.pgwire;
 import io.questdb.std.Unsafe;
 import org.junit.Assert;
 import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.junit.runners.Parameterized;
 import org.postgresql.util.PSQLException;
 
 import java.sql.PreparedStatement;
 import java.sql.ResultSet;
 import java.sql.Statement;
-import java.util.Collection;
 
 import static io.questdb.test.tools.TestUtils.assertContains;
 
-@RunWith(Parameterized.class)
 public class PGMemoryLimitTest extends BasePGTest {
-
-    public PGMemoryLimitTest(LegacyMode legacyMode) {
-        super(legacyMode);
-    }
-
-    @Parameterized.Parameters(name = "{0}")
-    public static Collection<Object[]> testParams() {
-        return legacyModeParams();
-    }
 
     @Test
     public void testUpdateRecoversFromOomError() throws Exception {
