+++ core/src/test/java/io/questdb/test/cutlass/pgwire/PgBootstrapTest.java
@@ -30,43 +30,20 @@ import io.questdb.test.tools.TestUtils;
 import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.junit.runners.Parameterized;
 import org.postgresql.util.PSQLException;
 
 import java.sql.Connection;
 import java.sql.DriverManager;
 import java.sql.SQLException;
-import java.util.Collection;
 import java.util.Properties;
 
-import static io.questdb.test.cutlass.pgwire.BasePGTest.LegacyMode.LEGACY;
-import static io.questdb.test.cutlass.pgwire.BasePGTest.legacyModeParams;
 
-@RunWith(Parameterized.class)
 public class PgBootstrapTest extends AbstractBootstrapTest {
-
-    private final boolean testParamLegacyMode;
-
-    public PgBootstrapTest(BasePGTest.LegacyMode legacyMode) {
-        this.testParamLegacyMode = legacyMode == LEGACY;
-    }
-
-    @Parameterized.Parameters(name = "{0}")
-    public static Collection<Object[]> testParams() {
-        return legacyModeParams();
-    }
 
     @Before
     public void setUp() {
         super.setUp();
-        TestUtils.unchecked(() -> {
-            if (testParamLegacyMode) {
-                createDummyConfiguration("pg.legacy.mode.enabled=true");
-            } else {
-                createDummyConfiguration();
-            }
-        });
+        TestUtils.unchecked(() -> createDummyConfiguration());
         dbPath.parent().$();
     }
